% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{adaptive_filter}
\alias{adaptive_filter}
\title{Adaptive filter}
\usage{
adaptive_filter(input, output, filterx = 11, filtery = 11,
  threshold = 2, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{filterx}{Size of the filter kernel in the x-direction.}

\item{filtery}{Size of the filter kernel in the y-direction.}

\item{threshold}{Difference from mean threshold, in standard deviations.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs an adaptive filter on an image.
}
