% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{add}
\alias{add}
\title{Add}
\usage{
add(input1, input2, output, verbose_mode = TRUE)
}
\arguments{
\item{input1}{Input raster file or constant value.}

\item{input2}{Input raster file or constant value.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs an addition operation on two rasters or a raster and a constant value.
}
