% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{attribute_correlation}
\alias{attribute_correlation}
\title{Attribute correlation}
\usage{
attribute_correlation(input, output = NULL, verbose_mode = TRUE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output HTML file (default name will be based on input file if unspecified).}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a correlation analysis on attribute fields from a vector database.
}
