% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{attribute_histogram}
\alias{attribute_histogram}
\title{Attribute histogram}
\usage{
attribute_histogram(input, field, output, verbose_mode = TRUE)
}
\arguments{
\item{input}{Input raster file.}

\item{field}{Input field name in attribute table.}

\item{output}{Output HTML file (default name will be based on input file if unspecified).}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Creates a histogram for the field values of a vector's attribute table.
}
