% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{average_flowpath_slope}
\alias{average_flowpath_slope}
\title{Average flowpath slope}
\usage{
average_flowpath_slope(dem, output, verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Measures the average slope gradient from each grid cell to all upslope divide cells.
}
