% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{average_normal_vector_angular_deviation}
\alias{average_normal_vector_angular_deviation}
\title{Average normal vector angular deviation}
\usage{
average_normal_vector_angular_deviation(dem, output, filter = 11,
  verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{filter}{Size of the filter kernel.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the circular variance of aspect at a scale for a DEM.
}
