% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{balance_contrast_enhancement}
\alias{balance_contrast_enhancement}
\title{Balance contrast enhancement}
\usage{
balance_contrast_enhancement(input, output, band_mean = 100,
  verbose_mode = TRUE)
}
\arguments{
\item{input}{Input colour composite image file.}

\item{output}{Output raster file.}

\item{band_mean}{Band mean value.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a balance contrast enhancement on a colour-composite image of multispectral data.
}
