% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breach_depressions.R
\name{breach_depressions}
\alias{breach_depressions}
\title{Title}
\usage{
breach_depressions(dem, output, max_depth = NULL, max_length = NULL)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{max_depth}{Optional maximum breach depth (default is Inf).}

\item{max_length}{Optional maximum breach channel length (in grid cells; default is Inf).}
}
\description{
Title
}
