% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{centroid_vector}
\alias{centroid_vector}
\title{Centroid vector}
\usage{
centroid_vector(input, output, verbose_mode = TRUE)
}
\arguments{
\item{input}{Input vector file.}

\item{output}{Output vector file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Identifes the centroid point of a vector polyline or polygon feature or a group of vector points.
}
