% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_whitebox_binary}
\alias{check_whitebox_binary}
\title{Check for WhiteboxTools Binary}
\usage{
check_whitebox_binary(silent = TRUE)
}
\arguments{
\item{silent}{logical. Print help on installation/setting binary path. Default \code{TRUE}.}
}
\value{
logical if WhiteboxTools Binary exists.
}
\description{
Check for WhiteboxTools Binary
}
\seealso{
\code{\link[=wbt_exe_path]{wbt_exe_path()}}
}
