% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{classify_overlap_points}
\alias{classify_overlap_points}
\title{Classify overlap points}
\usage{
classify_overlap_points(input, output, resolution = 2, filter = FALSE,
  verbose_mode = FALSE)
}
\arguments{
\item{input}{Input LiDAR file.}

\item{output}{Output LiDAR file.}

\item{resolution}{The size of the square area used to evaluate nearby points in the LiDAR data.}

\item{filter}{Filter out points from overlapping flightlines? If false, overlaps will simply be classified.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Classifies or filters LAS points in regions of overlapping flight lines.
}
