% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{clean_vector}
\alias{clean_vector}
\title{Clean vector}
\usage{
clean_vector(input, output, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input vector file.}

\item{output}{Output vector file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Removes null features and lines/polygons with fewer than the required number of vertices.
}
