% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{clip}
\alias{clip}
\title{Clip}
\usage{
clip(input, clip, output, verbose_mode = TRUE)
}
\arguments{
\item{input}{Input vector file.}

\item{clip}{Input clip polygon vector file.}

\item{output}{Output vector file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Extract all the features, or parts of features, that overlap with the features of the clip vector.
}
