% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{clump}
\alias{clump}
\title{Clump}
\usage{
clump(input, output, diag = TRUE, zero_back = FALSE,
  verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{diag}{Flag indicating whether diagonal connections should be considered.}

\item{zero_back}{Flag indicating whether zero values should be treated as a background.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Groups cells that form discrete areas, assigning them unique identifiers.
}
