% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{compactness_ratio}
\alias{compactness_ratio}
\title{Compactness ratio}
\usage{
compactness_ratio(input, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input vector polygon file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the compactness ratio (A/P), a measure of shape complexity, for vector polygons.
}
