% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{convert_nodata_to_zero}
\alias{convert_nodata_to_zero}
\title{Convert nodata to zero}
\usage{
convert_nodata_to_zero(input, output, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Converts nodata values in a raster to zero.
}
