% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{convert_raster_format}
\alias{convert_raster_format}
\title{Convert raster format}
\usage{
convert_raster_format(input, output, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Converts raster data from one format to another.
}
