% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{cost_allocation}
\alias{cost_allocation}
\title{Cost allocation}
\usage{
cost_allocation(source, backlink, output, verbose_mode = TRUE)
}
\arguments{
\item{source}{Input source raster file.}

\item{backlink}{Input backlink raster file generated by the cost-distance tool.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Identifies the source cell to which each grid cell is connected by a least-cost pathway in a cost-distance analysis.
}
