% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{cost_distance}
\alias{cost_distance}
\title{Cost distance}
\usage{
cost_distance(source, cost, out_accum, out_backlink, verbose_mode = TRUE)
}
\arguments{
\item{source}{Input source raster file.}

\item{cost}{Input cost (friction) raster file.}

\item{out_accum}{Output cost accumulation raster file.}

\item{out_backlink}{Output backlink raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs cost-distance accumulation on a cost surface and a group of source cells.
}
