% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{cost_pathway}
\alias{cost_pathway}
\title{Cost pathway}
\usage{
cost_pathway(destination, backlink, output, zero_background = FALSE,
  verbose_mode = FALSE)
}
\arguments{
\item{destination}{Input destination raster file.}

\item{backlink}{Input backlink raster file generated by the cost-distance tool.}

\item{output}{Output cost pathway raster file.}

\item{zero_background}{Flag indicating whether zero values should be treated as a background.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs cost-distance pathway analysis using a series of destination grid cells.
}
