% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{count_if}
\alias{count_if}
\title{Count if}
\usage{
count_if(inputs, output, value, verbose_mode = TRUE)
}
\arguments{
\item{inputs}{Input raster files.}

\item{output}{Output raster file.}

\item{value}{Search value (e.g. countif value = 5.0).}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Counts the number of occurrences of a specified value in a cell-stack of rasters.
}
