% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{create_plane}
\alias{create_plane}
\title{Create plane}
\usage{
create_plane(base, output, gradient = 15, aspect = 90, constant = 0,
  verbose_mode = TRUE)
}
\arguments{
\item{base}{Input base raster file.}

\item{output}{Output raster file.}

\item{gradient}{Slope gradient in degrees (-85.0 to 85.0).}

\item{aspect}{Aspect (direction) in degrees clockwise from north (0.0-360.0).}

\item{constant}{Constant value.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Creates a raster image based on the equation for a simple plane.
}
