% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{crispness_index}
\alias{crispness_index}
\title{Crispness index}
\usage{
crispness_index(input, output = NULL, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Optional output html file (default name will be based on input file if unspecified).}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the Crispness Index, which is used to quantify how crisp (or conversely how fuzzy) a probability image is.
}
