% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{d8_flow_accumulation}
\alias{d8_flow_accumulation}
\title{D8 flow accumulation}
\usage{
d8_flow_accumulation(dem, output, out_type = "cells", log = FALSE,
  clip = FALSE, verbose_mode = TRUE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{out_type}{Output type; one of 'cells' (default), 'catchment area', and 'specific contributing area'.}

\item{log}{Optional flag to request the output be log-transformed.}

\item{clip}{Optional flag to request clipping the display max by 1 percent.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates a D8 flow accumulation raster from an input DEM.
}
