% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{d_inf_flow_accumulation}
\alias{d_inf_flow_accumulation}
\title{D inf flow accumulation}
\usage{
d_inf_flow_accumulation(dem, output,
  out_type = "Specific Contributing Area", threshold = NULL,
  log = FALSE, clip = FALSE, verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{out_type}{Output type; one of 'cells', 'sca' (default), and 'ca'.}

\item{threshold}{Optional convergence threshold parameter, in grid cells; default is inifinity.}

\item{log}{Optional flag to request the output be log-transformed.}

\item{clip}{Optional flag to request clipping the display max by 1 percent.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates a D-infinity flow accumulation raster from an input DEM.
}
