% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{diff_from_mean_elev}
\alias{diff_from_mean_elev}
\title{Diff from mean elev}
\usage{
diff_from_mean_elev(dem, output, filterx = 11, filtery = 11,
  verbose_mode = TRUE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{filterx}{Size of the filter kernel in the x-direction.}

\item{filtery}{Size of the filter kernel in the y-direction.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates difference from mean elevation (equivalent to a high-pass filter).
}
