% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{direct_decorrelation_stretch}
\alias{direct_decorrelation_stretch}
\title{Direct decorrelation stretch}
\usage{
direct_decorrelation_stretch(input, output, k = 0.5, clip = 1,
  verbose_mode = TRUE)
}
\arguments{
\item{input}{Input colour composite image file.}

\item{output}{Output raster file.}

\item{k}{Achromatic factor (k) ranges between 0 (no effect) and 1 (full saturation stretch), although typical values range from 0.3 to 0.7.}

\item{clip}{Optional percent to clip the upper tail by during the stretch.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a direct decorrelation stretch enhancement on a colour-composite image of multispectral data.
}
