% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{downslope_distance_to_stream}
\alias{downslope_distance_to_stream}
\title{Downslope distance to stream}
\usage{
downslope_distance_to_stream(dem, streams, output, verbose_mode = TRUE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{streams}{Input raster streams file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Measures distance to the nearest downslope stream cell.
}
