% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{edge_preserving_mean_filter}
\alias{edge_preserving_mean_filter}
\title{Edge preserving mean filter}
\usage{
edge_preserving_mean_filter(input, output, threshold, filter = 11,
  verbose_mode = TRUE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{threshold}{Maximum difference in values.}

\item{filter}{Size of the filter kernel.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a simple edge-preserving mean filter on an input image.
}
