% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{edge_proportion}
\alias{edge_proportion}
\title{Edge proportion}
\usage{
edge_proportion(input, output, output_text = FALSE,
  verbose_mode = TRUE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{output_text}{flag indicating whether a text report should also be output.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculate the proportion of cells in a raster polygon that are edge cells.
}
