% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{elev_above_pit}
\alias{elev_above_pit}
\title{Elev above pit}
\usage{
elev_above_pit(dem, output, verbose_mode = TRUE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculate the elevation of each grid cell above the nearest downstream pit cell or grid edge cell.
}
