% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{elev_relative_to_min_max}
\alias{elev_relative_to_min_max}
\title{Elev relative to min max}
\usage{
elev_relative_to_min_max(dem, output, verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the elevation of a location relative to the minimum and maximum elevations in a DEM.
}
