% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{elevation_above_stream}
\alias{elevation_above_stream}
\title{Elevation above stream}
\usage{
elevation_above_stream(dem, streams, output, verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{streams}{Input raster streams file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the elevation of cells above the nearest downslope stream cell.
}
