% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{erase_polygon_from_raster}
\alias{erase_polygon_from_raster}
\title{Erase polygon from raster}
\usage{
erase_polygon_from_raster(input, polygons, output, verbose_mode = TRUE)
}
\arguments{
\item{input}{Input raster file.}

\item{polygons}{Input vector polygons file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Erases (cuts out) a vector polygon from a raster.
}
