% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{extract_nodes}
\alias{extract_nodes}
\title{Extract nodes}
\usage{
extract_nodes(input, output, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input vector lines or polygon file.}

\item{output}{Output vector points file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Converts vector lines or polygons into vertex points.
}
