% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{extract_raster_statistics}
\alias{extract_raster_statistics}
\title{Extract raster statistics}
\usage{
extract_raster_statistics(input, features, output = NULL,
  stat = "average", out_table = NULL, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input data raster file.}

\item{features}{Input feature definition raster file.}

\item{output}{Output raster file.}

\item{stat}{Statistic to extract.}

\item{out_table}{Output HTML Table file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Extracts descriptive statistics for a group of patches in a raster.
}
