% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{extract_raster_values_at_points}
\alias{extract_raster_values_at_points}
\title{Extract raster values at points}
\usage{
extract_raster_values_at_points(inputs, points, verbose_mode = FALSE)
}
\arguments{
\item{inputs}{Input raster files.}

\item{points}{Input vector points file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Extracts the values of raster(s) at vector point locations.
}
