% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_network_analysis.R
\name{extract_streams}
\alias{extract_streams}
\title{Extract streams}
\usage{
extract_streams(flow_accum, output, threshold, zero_background = FALSE,
  verbose_mode = FALSE)
}
\arguments{
\item{flow_accum}{Input raster D8 flow accumulation file.}

\item{output}{Output raster file.}

\item{threshold}{Threshold in flow accumulation values for channelization.}

\item{zero_background}{Flag indicating whether a background value of zero should be used.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Extracts stream grid cells from a flow accumulation raster.
}
