% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_network_analysis.R
\name{extract_valleys}
\alias{extract_valleys}
\title{Extract valleys}
\usage{
extract_valleys(dem, output, variant = "LQ", line_thin = TRUE,
  filter = 5, verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{variant}{Options include 'LQ' (lower quartile), 'JandR' (Johnston and Rosenfeld), and 'PandD' (Peucker and Douglas); default is 'LQ'.}

\item{line_thin}{Optional flag indicating whether post-processing line-thinning should be performed.}

\item{filter}{Optional argument (only used when variant='lq') providing the filter size, in grid cells, used for lq-filtering (default is 5).}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Identifies potential valley bottom grid cells based on local topolography alone.
}
