% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{fast_almost_gaussian_filter}
\alias{fast_almost_gaussian_filter}
\title{Fast almost gaussian filter}
\usage{
fast_almost_gaussian_filter(input, output, sigma = 1.8,
  verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{sigma}{Standard deviation distance in pixels.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a fast approximate Gaussian filter on an image.
}
