% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{fd8_flow_accumulation}
\alias{fd8_flow_accumulation}
\title{Fd8 flow accumulation}
\usage{
fd8_flow_accumulation(dem, output,
  out_type = "specific contributing area", exponent = 1.1,
  threshold = NULL, log = FALSE, clip = FALSE,
  verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{out_type}{Output type; one of 'cells', 'specific contributing area' (default), and 'catchment area'.}

\item{exponent}{Optional exponent parameter; default is 1.1.}

\item{threshold}{Optional convergence threshold parameter, in grid cells; default is inifinity.}

\item{log}{Optional flag to request the output be log-transformed.}

\item{clip}{Optional flag to request clipping the display max by 1 percent.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates an FD8 flow accumulation raster from an input DEM.
}
