% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{fetch_analysis}
\alias{fetch_analysis}
\title{Fetch analysis}
\usage{
fetch_analysis(dem, output, azimuth = 0, hgt_inc = 0.05,
  verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{azimuth}{Wind azimuth in degrees in degrees.}

\item{hgt_inc}{Height increment value.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs an analysis of fetch or upwind distance to an obstacle.
}
