% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{fill_burn}
\alias{fill_burn}
\title{Fill burn}
\usage{
fill_burn(dem, streams, output, verbose_mode = TRUE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{streams}{Input vector streams file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Burns streams into a DEM using the FillBurn (Saunders, 1999) method.
}
