% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{fill_missing_data}
\alias{fill_missing_data}
\title{Fill missing data}
\usage{
fill_missing_data(input, output, filter = 11, weight = 2,
  verbose_mode = TRUE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{filter}{Filter size (cells).}

\item{weight}{IDW weight value.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Fills nodata holes in a DEM.
}
