% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{find_lowest_or_highest_points}
\alias{find_lowest_or_highest_points}
\title{Find lowest or highest points}
\usage{
find_lowest_or_highest_points(input, output, out_type = "lowest",
  verbose_mode = TRUE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output vector points file.}

\item{out_type}{Output type; one of 'area' (default) and 'volume'.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Locates the lowest and/or highest valued cells in a raster.
}
