% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{find_no_flow_cells}
\alias{find_no_flow_cells}
\title{Find no flow cells}
\usage{
find_no_flow_cells(dem, output, verbose_mode = TRUE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Finds grid cells with no downslope neighbours.
}
