% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{find_patch_or_class_edge_cells}
\alias{find_patch_or_class_edge_cells}
\title{Find patch or class edge cells}
\usage{
find_patch_or_class_edge_cells(input, output, verbose_mode = TRUE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Finds all cells located on the edge of patch or class features.
}
