% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{find_ridges}
\alias{find_ridges}
\title{Find ridges}
\usage{
find_ridges(dem, output, line_thin = TRUE, verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{line_thin}{Optional flag indicating whether post-processing line-thinning should be performed.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Identifies potential ridge and peak grid cells.
}
