% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{flightline_overlap}
\alias{flightline_overlap}
\title{Flightline overlap}
\usage{
flightline_overlap(input, output = NULL, resolution = 1,
  verbose_mode = FALSE)
}
\arguments{
\item{input}{Input LiDAR file.}

\item{output}{Output file.}

\item{resolution}{Output raster's grid resolution.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Reads a LiDAR (LAS) point file and outputs a raster containing the number of overlapping flight lines in each grid cell.
}
