% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{flood_order}
\alias{flood_order}
\title{Flood order}
\usage{
flood_order(dem, output, verbose_mode = TRUE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Assigns each DEM grid cell its order in the sequence of inundations that are encountered during a search starting from the edges, moving inward at increasing elevations.
}
