% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{gamma_correction}
\alias{gamma_correction}
\title{Gamma correction}
\usage{
gamma_correction(input, output, gamma = 0.5, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{gamma}{Gamma value.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a gamma correction on an input images.
}
