% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbt.R
\name{get_whitebox_exe}
\alias{get_whitebox_exe}
\title{Get file path of WhiteboxTools executable.}
\usage{
get_whitebox_exe()
}
\value{
Returns the file path of WhiteboxTools executable.
}
\description{
Get file path of WhiteboxTools executable.
}
\examples{
wbt_exe <- get_whitebox_exe()
}
