% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{highest_position}
\alias{highest_position}
\title{Highest position}
\usage{
highest_position(inputs, output, verbose_mode = TRUE)
}
\arguments{
\item{inputs}{Input raster files.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Identifies the stack position of the maximum value within a raster stack on a cell-by-cell basis.
}
