% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{histogram_equalization}
\alias{histogram_equalization}
\title{Histogram equalization}
\usage{
histogram_equalization(input, output, num_tones = 256,
  verbose_mode = TRUE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{num_tones}{Number of tones in the output image.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a histogram equalization contrast enhancment on an image.
}
