% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{histogram_matching}
\alias{histogram_matching}
\title{Histogram matching}
\usage{
histogram_matching(input, histo_file, output, verbose_mode = TRUE)
}
\arguments{
\item{input}{Input raster file.}

\item{histo_file}{Input reference probability distribution function (pdf) text file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Alters the statistical distribution of a raster image matching it to a specified PDF.
}
