% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{image_autocorrelation}
\alias{image_autocorrelation}
\title{Image autocorrelation}
\usage{
image_autocorrelation(inputs, output, contiguity = "Rook",
  verbose_mode = TRUE)
}
\arguments{
\item{inputs}{Input raster files.}

\item{output}{Output HTML file (default name will be based on input file if unspecified).}

\item{contiguity}{Contiguity type.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs Moran's I analysis on two or more input images.
}
