% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{image_regression}
\alias{image_regression}
\title{Image regression}
\usage{
image_regression(input1, input2, output, out_residuals = NULL,
  standardize = FALSE, verbose_mode = FALSE)
}
\arguments{
\item{input1}{Input raster file (independent variable, X).}

\item{input2}{Input raster file (dependent variable, Y).}

\item{output}{Output HTML file for regression summary report.}

\item{out_residuals}{Output raster regression resdidual file.}

\item{standardize}{Optional flag indicating whether to standardize the residuals map.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs image regression analysis on two input images.
}
