% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{in_place_multiply}
\alias{in_place_multiply}
\title{In place multiply}
\usage{
in_place_multiply(input1, input2, verbose_mode = TRUE)
}
\arguments{
\item{input1}{Input raster file.}

\item{input2}{Input raster file or constant value.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs an in-place multiplication operation (input1 *= input2).
}
