% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbt.R
\name{install_whitebox}
\alias{install_whitebox}
\alias{wbt_install}
\title{Download and Install WhiteboxTools}
\usage{
install_whitebox(pkg_dir = find.package("whitebox"))
}
\arguments{
\item{pkg_dir}{default install path is to whitebox package "WBT" folder}
}
\value{
Prints out the location of the WhiteboxTools binary
}
\description{
This function downloads the WhiteboxTools binary if needed. This only works for 64-bit Linux, Windows and Mac OSX platforms. If you need WhiteboxTools for another platform follow the instructions here: \url{https://github.com/jblindsay/whitebox-tools}
}
\examples{
\dontrun{
install_whitebox()
}
}
