% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{integral_image}
\alias{integral_image}
\title{Integral image}
\usage{
integral_image(input, output, verbose_mode = TRUE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Transforms an input image (summed area table) into its integral image equivalent.
}
