% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{isobasins}
\alias{isobasins}
\title{Isobasins}
\usage{
isobasins(dem, output, size, verbose_mode = TRUE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{size}{Target basin size, in grid cells.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Divides a landscape into nearly equal sized drainage basins (i.e. watersheds).
}
