% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{jenson_snap_pour_points}
\alias{jenson_snap_pour_points}
\title{Jenson snap pour points}
\usage{
jenson_snap_pour_points(pour_pts, streams, output, snap_dist,
  verbose_mode = TRUE)
}
\arguments{
\item{pour_pts}{Input vector pour points (outlet) file.}

\item{streams}{Input raster streams file.}

\item{output}{Output vector file.}

\item{snap_dist}{Maximum snap distance in map units.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Moves outlet points used to specify points of interest in a watershedding operation to the nearest stream cell.
}
