% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{laplacian_of_gaussian_filter}
\alias{laplacian_of_gaussian_filter}
\title{Laplacian of gaussian filter}
\usage{
laplacian_of_gaussian_filter(input, output, sigma = 0.75,
  verbose_mode = TRUE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{sigma}{Standard deviation in pixels.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a Laplacian-of-Gaussian (LoG) filter on an image.
}
