% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{las_to_ascii}
\alias{las_to_ascii}
\title{Las to ascii}
\usage{
las_to_ascii(inputs, verbose_mode = TRUE)
}
\arguments{
\item{inputs}{Input LiDAR files.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Converts one or more LAS files into ASCII text files.
}
