% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{las_to_multipoint_shapefile}
\alias{las_to_multipoint_shapefile}
\title{Las to multipoint shapefile}
\usage{
las_to_multipoint_shapefile(input, verbose_mode = TRUE)
}
\arguments{
\item{input}{Input LiDAR file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Converts one or more LAS files into MultipointZ vector Shapefiles. When the input parameter is not specified, the tool grids all LAS files contained within the working directory.
}
