% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{las_to_shapefile}
\alias{las_to_shapefile}
\title{Las to shapefile}
\usage{
las_to_shapefile(input, verbose_mode = TRUE)
}
\arguments{
\item{input}{Input LiDAR file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Converts one or more LAS files into a vector Shapefile of POINT ShapeType.
}
