% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{layer_footprint}
\alias{layer_footprint}
\title{Layer footprint}
\usage{
layer_footprint(input, output, verbose_mode = TRUE)
}
\arguments{
\item{input}{Input raster or vector file.}

\item{output}{Output vector polygon file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Creates a vector polygon footprint of the area covered by a raster grid or vector layer.
}
