% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{less_than}
\alias{less_than}
\title{Less than}
\usage{
less_than(input1, input2, output, incl_equals = FALSE,
  verbose_mode = FALSE)
}
\arguments{
\item{input1}{Input raster file or constant value.}

\item{input2}{Input raster file or constant value.}

\item{output}{Output raster file.}

\item{incl_equals}{Perform a less-than-or-equal-to operation.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a less-than comparison operation on two rasters or a raster and a constant value.
}
