% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{lidar_block_maximum}
\alias{lidar_block_maximum}
\title{Lidar block maximum}
\usage{
lidar_block_maximum(input, output = NULL, resolution = 1,
  verbose_mode = TRUE)
}
\arguments{
\item{input}{Input LiDAR file.}

\item{output}{Output file.}

\item{resolution}{Output raster's grid resolution.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Creates a block-maximum raster from an input LAS file. When the input/output parameters are not specified, the tool grids all LAS files contained within the working directory.
}
