% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{lidar_classify_subset}
\alias{lidar_classify_subset}
\title{Lidar classify subset}
\usage{
lidar_classify_subset(base, subset, output, subset_class,
  nonsubset_class = NULL, verbose_mode = FALSE)
}
\arguments{
\item{base}{Input base LiDAR file.}

\item{subset}{Input subset LiDAR file.}

\item{output}{Output LiDAR file.}

\item{subset_class}{Subset point class value (must be 0-18; see LAS specifications).}

\item{nonsubset_class}{Non-subset point class value (must be 0-18; see LAS specifications).}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Classifies the values in one LiDAR point cloud that correpond with points in a subset cloud.
}
