% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{lidar_elevation_slice}
\alias{lidar_elevation_slice}
\title{Lidar elevation slice}
\usage{
lidar_elevation_slice(input, output, minz = NULL, maxz = NULL,
  cls = FALSE, inclassval = 2, outclassval = 1,
  verbose_mode = FALSE)
}
\arguments{
\item{input}{Input LiDAR file.}

\item{output}{Output LiDAR file.}

\item{minz}{Minimum elevation value (optional).}

\item{maxz}{Maximum elevation value (optional).}

\item{cls}{Optional boolean flag indicating whether points outside the range should be retained in output but reclassified.}

\item{inclassval}{Optional parameter specifying the class value assigned to points within the slice.}

\item{outclassval}{Optional parameter specifying the class value assigned to points within the slice.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Outputs all of the points within a LiDAR (LAS) point file that lie between a specified elevation range.
}
