% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{lidar_join}
\alias{lidar_join}
\title{Lidar join}
\usage{
lidar_join(inputs, output, verbose_mode = FALSE)
}
\arguments{
\item{inputs}{Input LiDAR files.}

\item{output}{Output LiDAR file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Joins multiple LiDAR (LAS) files into a single LAS file.
}
