% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{lidar_point_stats}
\alias{lidar_point_stats}
\title{Lidar point stats}
\usage{
lidar_point_stats(input, resolution = 1, num_points = TRUE,
  num_pulses = FALSE, z_range = FALSE, intensity_range = FALSE,
  predom_class = FALSE, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input LiDAR file.}

\item{resolution}{Output raster's grid resolution.}

\item{num_points}{Flag indicating whether or not to output the number of points raster.}

\item{num_pulses}{Flag indicating whether or not to output the number of pulses raster.}

\item{z_range}{Flag indicating whether or not to output the elevation range raster.}

\item{intensity_range}{Flag indicating whether or not to output the intensity range raster.}

\item{predom_class}{Flag indicating whether or not to output the predominant classification raster.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Creates several rasters summarizing the distribution of LAS point data. When the input/output parameters are not specified, the tool works on all LAS files contained within the working directory.
}
