% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{lidar_remove_duplicates}
\alias{lidar_remove_duplicates}
\title{Lidar remove duplicates}
\usage{
lidar_remove_duplicates(input, output, include_z = FALSE,
  verbose_mode = FALSE)
}
\arguments{
\item{input}{Input LiDAR file.}

\item{output}{Output LiDAR file.}

\item{include_z}{Include z-values in point comparison?.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Removes duplicate points from a LiDAR data set.
}
