% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{lidar_remove_outliers}
\alias{lidar_remove_outliers}
\title{Lidar remove outliers}
\usage{
lidar_remove_outliers(input, output, radius = 2, elev_diff = 50,
  verbose_mode = FALSE)
}
\arguments{
\item{input}{Input LiDAR file.}

\item{output}{Output LiDAR file.}

\item{radius}{Search Radius.}

\item{elev_diff}{Max. elevation difference.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Removes outliers (high and low points) in a LiDAR point cloud.
}
