% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{lidar_tile}
\alias{lidar_tile}
\title{Lidar tile}
\usage{
lidar_tile(input, width_x = 1000, width_y = 1000, origin_x = 0,
  origin_y = 0, min_points = 2, verbose_mode = TRUE)
}
\arguments{
\item{input}{Input LiDAR file.}

\item{width_x}{Width of tiles in the X dimension; default 1000.0.}

\item{width_y}{Width of tiles in the Y dimension.}

\item{origin_x}{Origin point X coordinate for tile grid.}

\item{origin_y}{Origin point Y coordinate for tile grid.}

\item{min_points}{Minimum number of points contained in a tile for it to be saved.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Tiles a LiDAR LAS file into multiple LAS files.
}
