% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{lidar_tile_footprint}
\alias{lidar_tile_footprint}
\title{Lidar tile footprint}
\usage{
lidar_tile_footprint(input, output, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input LiDAR file.}

\item{output}{Output vector polygon file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Creates a vector polygon of the convex hull of a LiDAR point cloud. When the input/output parameters are not specified, the tool works with all LAS files contained within the working directory.
}
