% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{line_detection_filter}
\alias{line_detection_filter}
\title{Line detection filter}
\usage{
line_detection_filter(input, output, variant = "vertical",
  absvals = FALSE, clip = 0, verbose_mode = TRUE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{variant}{Optional variant value. Options include 'v' (vertical), 'h' (horizontal), '45', and '135' (default is 'v').}

\item{absvals}{Optional flag indicating whether outputs should be absolute values.}

\item{clip}{Optional amount to clip the distribution tails by, in percent.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a line-detection filter on an image.
}
