% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{line_thinning}
\alias{line_thinning}
\title{Line thinning}
\usage{
line_thinning(input, output, verbose_mode = TRUE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs line thinning a on Boolean raster image; intended to be used with the RemoveSpurs tool.
}
