% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{lines_to_polygons}
\alias{lines_to_polygons}
\title{Lines to polygons}
\usage{
lines_to_polygons(input, output, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input vector line file.}

\item{output}{Output vector polygon file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Converts vector polylines to polygons.
}
