% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{log2}
\alias{log2}
\title{Log2}
\usage{
log2(input, output, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Returns the base-2 logarithm of values in a raster.
}
\examples{
dem <- system.file("extdata", "DEM.tif", package="whitebox")
log2(input = dem, output = 'output.tif')
}
