% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_network_analysis.R
\name{long_profile}
\alias{long_profile}
\title{Long profile}
\usage{
long_profile(d8_pntr, streams, dem, output, esri_pntr = FALSE,
  verbose_mode = FALSE)
}
\arguments{
\item{d8_pntr}{Input raster D8 pointer file.}

\item{streams}{Input raster streams file.}

\item{dem}{Input raster DEM file.}

\item{output}{Output HTML file.}

\item{esri_pntr}{D8 pointer uses the ESRI style scheme.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Plots the stream longitudinal profiles for one or more rivers.
}
