% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{longest_flowpath}
\alias{longest_flowpath}
\title{Longest flowpath}
\usage{
longest_flowpath(dem, basins, output, verbose_mode = TRUE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{basins}{Input raster basins file.}

\item{output}{Output vector file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Delineates the longest flowpaths for a group of subbasins or watersheds.
}
