% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{lowest_position}
\alias{lowest_position}
\title{Lowest position}
\usage{
lowest_position(inputs, output, verbose_mode = FALSE)
}
\arguments{
\item{inputs}{Input raster files.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Identifies the stack position of the minimum value within a raster stack on a cell-by-cell basis.
}
