% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{max_branch_length}
\alias{max_branch_length}
\title{Max branch length}
\usage{
max_branch_length(dem, output, log = FALSE, verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{log}{Optional flag to request the output be log-transformed.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Lindsay and Seibert's (2013) branch length index is used to map drainage divides or ridge lines.
}
