% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{max_upslope_flowpath_length}
\alias{max_upslope_flowpath_length}
\title{Max upslope flowpath length}
\usage{
max_upslope_flowpath_length(dem, output, verbose_mode = TRUE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Measures the maximum length of all upslope flowpaths draining each grid cell.
}
