% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{mean_filter}
\alias{mean_filter}
\title{Mean filter}
\usage{
mean_filter(input, output, filterx = 3, filtery = 3,
  verbose_mode = TRUE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{filterx}{Size of the filter kernel in the x-direction.}

\item{filtery}{Size of the filter kernel in the y-direction.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a mean filter (low-pass filter) on an input image.
}
