% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{medoid}
\alias{medoid}
\title{Medoid}
\usage{
medoid(input, output, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input vector file.}

\item{output}{Output vector file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the medoid for a series of vector features contained in a shapefile.
}
