% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{merge_table_with_csv}
\alias{merge_table_with_csv}
\title{Merge table with csv}
\usage{
merge_table_with_csv(input, pkey, csv, fkey, import_field = NULL,
  verbose_mode = FALSE)
}
\arguments{
\item{input}{Input primary vector file (i.e. the table to be modified).}

\item{pkey}{Primary key field.}

\item{csv}{Input CSV file (i.e. source of data to be imported).}

\item{fkey}{Foreign key field.}

\item{import_field}{Imported field (all fields will be imported if not specified).}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Merge a vector's attribute table with a table contained within a CSV text file.
}
