% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{min_max_contrast_stretch}
\alias{min_max_contrast_stretch}
\title{Min max contrast stretch}
\usage{
min_max_contrast_stretch(input, output, min_val, max_val,
  num_tones = 256, verbose_mode = TRUE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{min_val}{Lower tail clip value.}

\item{max_val}{Upper tail clip value.}

\item{num_tones}{Number of tones in the output image.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a min-max contrast stretch on an input greytone image.
}
