% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{min_overlay}
\alias{min_overlay}
\title{Min overlay}
\usage{
min_overlay(inputs, output, verbose_mode = TRUE)
}
\arguments{
\item{inputs}{Input raster files.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Evaluates the minimum value for each grid cell from a stack of input rasters.
}
