% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{minimum_bounding_envelope}
\alias{minimum_bounding_envelope}
\title{Minimum bounding envelope}
\usage{
minimum_bounding_envelope(input, output, features = TRUE,
  verbose_mode = FALSE)
}
\arguments{
\item{input}{Input vector file.}

\item{output}{Output vector polygon file.}

\item{features}{Find the minimum bounding envelop around each individual vector feature.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Creates a vector axis-aligned minimum bounding rectangle (envelope) around vector features.
}
