% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{minimum_convex_hull}
\alias{minimum_convex_hull}
\title{Minimum convex hull}
\usage{
minimum_convex_hull(input, output, features = TRUE,
  verbose_mode = TRUE)
}
\arguments{
\item{input}{Input vector file.}

\item{output}{Output vector polygon file.}

\item{features}{Find the hulls around each vector feature.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Creates a vector convex polygon around vector features.
}
