% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{modified_k_means_clustering}
\alias{modified_k_means_clustering}
\title{Modified k means clustering}
\usage{
modified_k_means_clustering(inputs, output, out_html = NULL,
  start_clusters = 1000, merger_dist = NULL, max_iterations = 10,
  class_change = 2, verbose_mode = TRUE)
}
\arguments{
\item{inputs}{Input raster files.}

\item{output}{Output raster file.}

\item{out_html}{Output HTML report file.}

\item{start_clusters}{Initial number of clusters.}

\item{merger_dist}{Cluster merger distance.}

\item{max_iterations}{Maximum number of iterations.}

\item{class_change}{Minimum percent of cells changed between iterations before completion.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a modified k-means clustering operation on a multi-spectral dataset.
}
