% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{normal_vectors}
\alias{normal_vectors}
\title{Normal vectors}
\usage{
normal_vectors(input, output, radius = 1, verbose_mode = TRUE)
}
\arguments{
\item{input}{Input LiDAR file.}

\item{output}{Output LiDAR file.}

\item{radius}{Search Radius.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates normal vectors for points within a LAS file and stores these data (XYZ vector components) in the RGB field.
}
