% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{normalized_difference_vegetation_index}
\alias{normalized_difference_vegetation_index}
\title{Normalized difference vegetation index}
\usage{
normalized_difference_vegetation_index(nir, red, output, clip = 0,
  osavi = FALSE, verbose_mode = TRUE)
}
\arguments{
\item{nir}{Input near-infrared band image.}

\item{red}{Input red band image.}

\item{output}{Output raster file.}

\item{clip}{Optional amount to clip the distribution tails by, in percent.}

\item{osavi}{Optional flag indicating whether the optimized soil-adjusted veg index (OSAVI) should be used.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the normalized difference vegetation index (NDVI) from near-infrared and red imagery.
}
