% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{num_downslope_neighbours}
\alias{num_downslope_neighbours}
\title{Num downslope neighbours}
\usage{
num_downslope_neighbours(dem, output, verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the number of downslope neighbours to each grid cell in a DEM.
}
