% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{num_inflowing_neighbours}
\alias{num_inflowing_neighbours}
\title{Num inflowing neighbours}
\usage{
num_inflowing_neighbours(dem, output, verbose_mode = TRUE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Computes the number of inflowing neighbours to each cell in an input DEM based on the D8 algorithm.
}
