% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{opening}
\alias{opening}
\title{Opening}
\usage{
opening(input, output, filterx = 11, filtery = 11,
  verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{filterx}{Size of the filter kernel in the x-direction.}

\item{filtery}{Size of the filter kernel in the y-direction.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
An opening is a mathematical morphology operation involving a dilation (max filter) of an erosion (min filter) set.
}
