% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{percent_elev_range}
\alias{percent_elev_range}
\title{Percent elev range}
\usage{
percent_elev_range(dem, output, filterx = 3, filtery = 3,
  verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{filterx}{Size of the filter kernel in the x-direction.}

\item{filtery}{Size of the filter kernel in the y-direction.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates percent of elevation range from a DEM.
}
