% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{polygon_perimeter}
\alias{polygon_perimeter}
\title{Polygon perimeter}
\usage{
polygon_perimeter(input, verbose_mode = TRUE)
}
\arguments{
\item{input}{Input vector polygon file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the perimeter of vector polygons.
}
