% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{polygonize}
\alias{polygonize}
\title{Polygonize}
\usage{
polygonize(inputs, output, verbose_mode = FALSE)
}
\arguments{
\item{inputs}{Input vector polyline file.}

\item{output}{Output vector polygon file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Creates a polygon layer from two or more intersecting line features contained in one or more input vector line files.
}
