% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{power}
\alias{power}
\title{Power}
\usage{
power(input1, input2, output, verbose_mode = FALSE)
}
\arguments{
\item{input1}{Input raster file or constant value.}

\item{input2}{Input raster file or constant value.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Raises the values in grid cells of one rasters, or a constant value, by values in another raster or constant value.
}
