% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{profile}
\alias{profile}
\title{Profile}
\usage{
profile(lines, surface, output, verbose_mode = FALSE)
}
\arguments{
\item{lines}{Input vector line file.}

\item{surface}{Input raster surface file.}

\item{output}{Output HTML file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Plots profiles from digital surface models.
}
