% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{radius_of_gyration}
\alias{radius_of_gyration}
\title{Radius of gyration}
\usage{
radius_of_gyration(input, output, text_output = FALSE,
  verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{text_output}{Optional text output.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the distance of cells from their polygon's centroid.
}
