% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{range_filter}
\alias{range_filter}
\title{Range filter}
\usage{
range_filter(input, output, filterx = 11, filtery = 11,
  verbose_mode = TRUE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{filterx}{Size of the filter kernel in the x-direction.}

\item{filtery}{Size of the filter kernel in the y-direction.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Assigns each cell in the output grid the range of values in a moving window centred on each grid cell in the input raster.
}
