% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{raster_cell_assignment}
\alias{raster_cell_assignment}
\title{Raster cell assignment}
\usage{
raster_cell_assignment(input, output, assign = "column",
  verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{assign}{Which variable would you like to assign to grid cells? Options include 'column', 'row', 'x', and 'y'.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Assign row or column number to cells.
}
