% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{raster_summary_stats}
\alias{raster_summary_stats}
\title{Raster summary stats}
\usage{
raster_summary_stats(input, verbose_mode = TRUE)
}
\arguments{
\item{input}{Input raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Measures a rasters average, standard deviation, num. non-nodata cells, and total.
}
