% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{raster_to_vector_lines}
\alias{raster_to_vector_lines}
\title{Raster to vector lines}
\usage{
raster_to_vector_lines(input, output, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster lines file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Converts a raster lines features into a vector of the POLYLINE shapetype.
}
