% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{raster_to_vector_points}
\alias{raster_to_vector_points}
\title{Raster to vector points}
\usage{
raster_to_vector_points(input, output, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output vector points file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Converts a raster dataset to a vector of the POINT shapetype.
}
