% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_network_analysis.R
\name{rasterize_streams}
\alias{rasterize_streams}
\title{Rasterize streams}
\usage{
rasterize_streams(streams, base, output, nodata = TRUE,
  feature_id = FALSE, verbose_mode = TRUE)
}
\arguments{
\item{streams}{Input vector streams file.}

\item{base}{Input base raster file.}

\item{output}{Output raster file.}

\item{nodata}{Use NoData value for background?.}

\item{feature_id}{Use feature number as output value?.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Rasterizes vector streams based on Lindsay (2016) method.
}
