% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{reclass_from_file}
\alias{reclass_from_file}
\title{Reclass from file}
\usage{
reclass_from_file(input, reclass_file, output, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{reclass_file}{Input text file containing reclass ranges.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Reclassifies the values in a raster image using reclass ranges in a text file.
}
