% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{reinitialize_attribute_table}
\alias{reinitialize_attribute_table}
\title{Reinitialize attribute table}
\usage{
reinitialize_attribute_table(input, verbose_mode = TRUE)
}
\arguments{
\item{input}{Input vector file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Reinitializes a vector's attribute table deleting all fields but the feature ID (FID).
}
