% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{related_circumscribing_circle}
\alias{related_circumscribing_circle}
\title{Related circumscribing circle}
\usage{
related_circumscribing_circle(input, verbose_mode = TRUE)
}
\arguments{
\item{input}{Input vector polygon file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the related circumscribing circle of vector polygons.
}
