% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{remove_spurs}
\alias{remove_spurs}
\title{Remove spurs}
\usage{
remove_spurs(input, output, iterations = 10, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{iterations}{Maximum number of iterations.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Removes the spurs (pruning operation) from a Boolean line image.; intended to be used on the output of the LineThinning tool.
}
