% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{rho8_pointer}
\alias{rho8_pointer}
\title{Rho8 pointer}
\usage{
rho8_pointer(dem, output, esri_pntr = FALSE, verbose_mode = TRUE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{esri_pntr}{D8 pointer uses the ESRI style scheme.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates a stochastic Rho8 flow pointer raster from an input DEM.
}
