% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{shape_complexity_index}
\alias{shape_complexity_index}
\title{Shape complexity index}
\usage{
shape_complexity_index(input, verbose_mode = TRUE)
}
\arguments{
\item{input}{Input vector polygon file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates overall polygon shape complexity or irregularity.
}
