% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{sink}
\alias{sink}
\title{Sink}
\usage{
sink(dem, output, zero_background = FALSE, verbose_mode = TRUE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{zero_background}{Flag indicating whether a background value of zero should be used.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Identifies the depressions in a DEM, giving each feature a unique identifier.
}
