% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{sobel_filter}
\alias{sobel_filter}
\title{Sobel filter}
\usage{
sobel_filter(input, output, variant = "3x3", clip = 0,
  verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{variant}{Optional variant value. Options include 3x3 and 5x5 (default is 3x3).}

\item{clip}{Optional amount to clip the distribution tails by, in percent (default is 0.0).}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a Sobel edge-detection filter on an image.
}
