% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{split_colour_composite}
\alias{split_colour_composite}
\title{Split colour composite}
\usage{
split_colour_composite(input, output, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input colour composite image file.}

\item{output}{Output raster file (suffixes of _r, _g, and _b will be appended).}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
This tool splits an RGB colour composite image into seperate multispectral images.
}
