% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{standard_deviation_contrast_stretch}
\alias{standard_deviation_contrast_stretch}
\title{Standard deviation contrast stretch}
\usage{
standard_deviation_contrast_stretch(input, output, stdev = 2,
  num_tones = 256, verbose_mode = TRUE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{stdev}{Standard deviation clip value.}

\item{num_tones}{Number of tones in the output image.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a standard-deviation contrast stretch on input images.
}
