% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{stochastic_depression_analysis}
\alias{stochastic_depression_analysis}
\title{Stochastic depression analysis}
\usage{
stochastic_depression_analysis(dem, output, rmse, range,
  iterations = 1000, verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output file.}

\item{rmse}{The DEM's root-mean-square-error (RMSE), in z units. This determines error magnitude.}

\item{range}{The error field's correlation length, in xy-units.}

\item{iterations}{The number of iterations.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Preforms a stochastic analysis of depressions within a DEM.
}
