% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_network_analysis.R
\name{stream_link_slope}
\alias{stream_link_slope}
\title{Stream link slope}
\usage{
stream_link_slope(d8_pntr, linkid, dem, output, esri_pntr = FALSE,
  zero_background = FALSE, verbose_mode = FALSE)
}
\arguments{
\item{d8_pntr}{Input raster D8 pointer file.}

\item{linkid}{Input raster streams link ID (or tributary ID) file.}

\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{esri_pntr}{D8 pointer uses the ESRI style scheme.}

\item{zero_background}{Flag indicating whether a background value of zero should be used.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Estimates the average slope of each link (or tributary) in a stream network.
}
