% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{subbasins}
\alias{subbasins}
\title{Subbasins}
\usage{
subbasins(d8_pntr, streams, output, esri_pntr = FALSE,
  verbose_mode = FALSE)
}
\arguments{
\item{d8_pntr}{Input D8 pointer raster file.}

\item{streams}{Input raster streams file.}

\item{output}{Output raster file.}

\item{esri_pntr}{D8 pointer uses the ESRI style scheme.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Identifies the catchments, or sub-basin, draining to each link in a stream network.
}
