% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{tan}
\alias{tan}
\title{Tan}
\usage{
tan(input, output, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Returns the tangent (tan) of each values in a raster.
}
\examples{
dem <- system.file("extdata", "DEM.tif", package="whitebox")
tan(input = dem, output = 'output.tif')
}
