% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{tophat_transform}
\alias{tophat_transform}
\title{Tophat transform}
\usage{
tophat_transform(input, output, filterx = 11, filtery = 11,
  variant = "white", verbose_mode = TRUE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{filterx}{Size of the filter kernel in the x-direction.}

\item{filtery}{Size of the filter kernel in the y-direction.}

\item{variant}{Optional variant value. Options include 'white' and 'black'.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs either a white or black top-hat transform on an input image.
}
