% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{trend_surface}
\alias{trend_surface}
\title{Trend surface}
\usage{
trend_surface(input, output, order = 1, verbose_mode = TRUE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{order}{Polynomial order (1 to 10).}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Estimates the trend surface of an input raster file.
}
