% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{turning_bands_simulation}
\alias{turning_bands_simulation}
\title{Turning bands simulation}
\usage{
turning_bands_simulation(base, output, range, iterations = 1000,
  verbose_mode = TRUE)
}
\arguments{
\item{base}{Input base raster file.}

\item{output}{Output file.}

\item{range}{The field's range, in xy-units, related to the extent of spatial autocorrelation.}

\item{iterations}{The number of iterations.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Creates an image containing random values based on a turning-bands simulation.
}
