% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{unnest_basins}
\alias{unnest_basins}
\title{Unnest basins}
\usage{
unnest_basins(d8_pntr, pour_pts, output, esri_pntr = FALSE,
  verbose_mode = TRUE)
}
\arguments{
\item{d8_pntr}{Input D8 pointer raster file.}

\item{pour_pts}{Input vector pour points (outlet) file.}

\item{output}{Output raster file.}

\item{esri_pntr}{D8 pointer uses the ESRI style scheme.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Extract whole watersheds for a set of outlet points.
}
