% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{vector_lines_to_raster}
\alias{vector_lines_to_raster}
\title{Vector lines to raster}
\usage{
vector_lines_to_raster(input, output, field = "FID", nodata = TRUE,
  cell_size = NULL, base = NULL, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input vector lines file.}

\item{output}{Output raster file.}

\item{field}{Input field name in attribute table.}

\item{nodata}{Background value to set to NoData. Without this flag, it will be set to 0.0.}

\item{cell_size}{Optionally specified cell size of output raster. Not used when base raster is specified.}

\item{base}{Optionally specified input base raster file. Not used when a cell size is specified.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Converts a vector containing polylines into a raster.
}
