% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{viewshed}
\alias{viewshed}
\title{Viewshed}
\usage{
viewshed(dem, stations, output, height = 2, verbose_mode = TRUE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{stations}{Input viewing station vector file.}

\item{output}{Output raster file.}

\item{height}{Viewing station height, in z units.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Identifies the viewshed for a point or set of points.
}
