% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{voronoi_diagram}
\alias{voronoi_diagram}
\title{Voronoi diagram}
\usage{
voronoi_diagram(input, output, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input vector points file.}

\item{output}{Output vector polygon file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Creates a vector Voronoi diagram for a set of vector points.
}
