% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{wbt_absolute_value}
\alias{wbt_absolute_value}
\title{Absolute value}
\usage{
wbt_absolute_value(input, output, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the absolute value of every cell in a raster.
}
\examples{
\dontrun{
dem <- system.file("extdata", "DEM.tif", package="whitebox")
wbt_absolute_value(input = dem, output = 'output.tif')
}
}
