% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_accumulation_curvature}
\alias{wbt_accumulation_curvature}
\title{Accumulation curvature}
\usage{
wbt_accumulation_curvature(
  dem,
  output,
  log = FALSE,
  zfactor = 1,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{dem}{Name of the input raster DEM file.}

\item{output}{Name of the output raster image file.}

\item{log}{Display output values using a log-scale.}

\item{zfactor}{Z conversion factor.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
This tool calculates accumulation curvature from an input DEM.
}
\keyword{GeomorphometricAnalysis}
