% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbt.R
\name{wbt_activate}
\alias{wbt_activate}
\title{Activate WhiteboxTools Extension Products}
\usage{
wbt_activate(
  email,
  activation_key,
  seat = 1,
  destdir = dirname(wbt_exe_path(shell_quote = FALSE))
)
}
\arguments{
\item{email}{Email Address}

\item{activation_key}{Activation Key}

\item{seat}{Seat Number (Default \code{1})}

\item{destdir}{Directory containing \code{whitebox_tools} and \verb{/plugins/} folder.}
}
\value{
\code{0} for success (invisibly). Try-error on error.
}
\description{
Activate WhiteboxTools Extension Products
}
