% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{wbt_add_point_coordinates_to_table}
\alias{wbt_add_point_coordinates_to_table}
\title{Add point coordinates to table}
\usage{
wbt_add_point_coordinates_to_table(input, wd = NULL, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input vector Points file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Modifies the attribute table of a point vector by adding fields containing each point's X and Y coordinates.
}
