% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_aggregate_raster}
\alias{wbt_aggregate_raster}
\title{Aggregate raster}
\usage{
wbt_aggregate_raster(
  input,
  output,
  agg_factor = 2,
  type = "mean",
  wd = NULL,
  verbose_mode = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{agg_factor}{Aggregation factor, in pixels.}

\item{type}{Statistic used to fill output pixels.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Aggregates a raster to a lower resolution.
}
