% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_ascii_to_las}
\alias{wbt_ascii_to_las}
\title{Ascii to las}
\usage{
wbt_ascii_to_las(inputs, pattern, proj = NULL, wd = NULL, verbose_mode = FALSE)
}
\arguments{
\item{inputs}{Input LiDAR  ASCII files (.csv).}

\item{pattern}{Input field pattern.}

\item{proj}{Well-known-text string or EPSG code describing projection.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Converts one or more ASCII files containing LiDAR points into LAS files.
}
