% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{wbt_attribute_correlation}
\alias{wbt_attribute_correlation}
\title{Attribute correlation}
\usage{
wbt_attribute_correlation(
  input,
  output = NULL,
  wd = NULL,
  verbose_mode = FALSE
)
}
\arguments{
\item{input}{Input vector file.}

\item{output}{Output HTML file (default name will be based on input file if unspecified).}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a correlation analysis on attribute fields from a vector database.
}
