% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis_enhancement.R
\name{wbt_balance_contrast_enhancement}
\alias{wbt_balance_contrast_enhancement}
\title{Balance contrast enhancement}
\usage{
wbt_balance_contrast_enhancement(
  input,
  output,
  band_mean = 100,
  wd = NULL,
  verbose_mode = NULL,
  compress_rasters = NULL,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input colour composite image file.}

\item{output}{Output raster file.}

\item{band_mean}{Band mean value.}

\item{wd}{Changes the working directory. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_wd()} for details.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_verbose()} for details.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_compress_rasters()} for details.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool. Default: \code{FALSE}.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a balance contrast enhancement on a colour-composite image of multispectral data.
}
\keyword{ImageProcessingToolsImageEnhancement}
