% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_boundary_shape_complexity}
\alias{wbt_boundary_shape_complexity}
\title{Boundary shape complexity}
\usage{
wbt_boundary_shape_complexity(input, output, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the complexity of the boundaries of raster polygons.
}
