% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{wbt_breach_depressions_least_cost}
\alias{wbt_breach_depressions_least_cost}
\title{Breach depressions least cost}
\usage{
wbt_breach_depressions_least_cost(dem, output, radius, max_cost = NULL,
  min_dist = TRUE, flat_increment = NULL, fill = TRUE,
  verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{radius}{.}

\item{max_cost}{Optional maximum breach cost (default is Inf).}

\item{min_dist}{Optional flag indicating whether to minimize breach distances.}

\item{flat_increment}{Optional elevation increment applied to flat areas.}

\item{fill}{Optional flag indicating whether to fill any remaining unbreached depressions.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Breaches the depressions in a DEM using a least-cost pathway method.
}
