% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis_distance.R
\name{wbt_buffer_raster}
\alias{wbt_buffer_raster}
\title{Buffer raster}
\usage{
wbt_buffer_raster(
  input,
  output,
  size,
  gridcells = FALSE,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{size}{Buffer size.}

\item{gridcells}{Optional flag to indicate that the 'size' threshold should be measured in grid cells instead of the default map units.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Maps a distance-based buffer around each non-background (non-zero/non-nodata) grid cell in an input image.
}
\keyword{GISAnalysisDistanceTools}
