% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis_filters.R
\name{wbt_canny_edge_detection}
\alias{wbt_canny_edge_detection}
\title{Canny edge detection
This tool performs a Canny edge-detection filter on an input image.}
\usage{
wbt_canny_edge_detection(
  input,
  output,
  sigma = 0.5,
  low = 0.05,
  high = 0.15,
  add_back = FALSE,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Name of the input raster image file.}

\item{output}{Name of the output raster image file.}

\item{sigma}{Sigma value used in Gaussian filtering, default = 0.5.}

\item{low}{Low threshold, default = 0.05.}

\item{high}{High threshold, default = 0.15.}

\item{add_back}{Add the edge cells back to the input image.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Canny edge detection
This tool performs a Canny edge-detection filter on an input image.
}
\keyword{ImageProcessingToolsFilters}
