% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{wbt_change_in_contributing_area}
\alias{wbt_change_in_contributing_area}
\title{Change in contributing area}
\usage{
wbt_change_in_contributing_area(
  dem,
  output,
  exponent = 1,
  threshold = NULL,
  log = FALSE,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{dem}{Name of the input DEM raster file; must be depressionless.}

\item{output}{Name of the output raster file.}

\item{exponent}{Optional exponent parameter; default is 1.0.}

\item{threshold}{Optional convergence threshold parameter, in grid cells; default is inifinity.}

\item{log}{Log-transform the output values?.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
This tool calculates the downslope rate of change in specific contributing area (SCA).
}
