% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{wbt_change_vector_analysis}
\alias{wbt_change_vector_analysis}
\title{Change vector analysis}
\usage{
wbt_change_vector_analysis(
  date1,
  date2,
  magnitude,
  direction,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{date1}{Input raster files for the earlier date.}

\item{date2}{Input raster files for the later date.}

\item{magnitude}{Output vector magnitude raster file.}

\item{direction}{Output vector Direction raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a change vector analysis on a two-date multi-spectral dataset.
}
