% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_clip_lidar_to_polygon}
\alias{wbt_clip_lidar_to_polygon}
\title{Clip lidar to polygon}
\usage{
wbt_clip_lidar_to_polygon(
  input,
  polygons,
  output,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{input}{Input LiDAR file.}

\item{polygons}{Input vector polygons file.}

\item{output}{Output LiDAR file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
Clips a LiDAR point cloud to a vector polygon or polygons.
}
