% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_colourize_based_on_class}
\alias{wbt_colourize_based_on_class}
\title{Colourize based on class}
\usage{
wbt_colourize_based_on_class(
  input,
  output = NULL,
  intensity_blending = 50,
  clr_str = "",
  use_unique_clrs_for_buildings = FALSE,
  radius = "",
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Name of the input LiDAR points.}

\item{output}{Name of the output LiDAR points.}

\item{intensity_blending}{Intensity blending amount (0-100 percent).}

\item{clr_str}{Colour values, e.g. 2: (184, 167, 108); 5: #9ab86c.}

\item{use_unique_clrs_for_buildings}{Use unique colours for each building?.}

\item{radius}{Search distance used in neighbourhood search.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Sets the RGB values of a LiDAR point cloud based on the point classification values.
}
