% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_cost_pathway}
\alias{wbt_cost_pathway}
\title{Cost pathway}
\usage{
wbt_cost_pathway(
  destination,
  backlink,
  output,
  zero_background = FALSE,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{destination}{Input destination raster file.}

\item{backlink}{Input backlink raster file generated by the cost-distance tool.}

\item{output}{Output cost pathway raster file.}

\item{zero_background}{Flag indicating whether zero values should be treated as a background.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs cost-distance pathway analysis using a series of destination grid cells.
}
