% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_create_hexagonal_vector_grid}
\alias{wbt_create_hexagonal_vector_grid}
\title{Create hexagonal vector grid}
\usage{
wbt_create_hexagonal_vector_grid(
  input,
  output,
  width,
  orientation = "horizontal",
  verbose_mode = FALSE
)
}
\arguments{
\item{input}{Input base file.}

\item{output}{Output vector polygon file.}

\item{width}{The grid cell width.}

\item{orientation}{Grid Orientation, 'horizontal' or 'vertical'.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Creates a hexagonal vector grid.
}
