% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_create_plane}
\alias{wbt_create_plane}
\title{Create plane}
\usage{
wbt_create_plane(
  base,
  output,
  gradient = 15,
  aspect = 90,
  constant = 0,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{base}{Input base raster file.}

\item{output}{Output raster file.}

\item{gradient}{Slope gradient in degrees (-85.0 to 85.0).}

\item{aspect}{Aspect (direction) in degrees clockwise from north (0.0-360.0).}

\item{constant}{Constant value.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
Creates a raster image based on the equation for a simple plane.
}
