% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_create_rectangular_vector_grid}
\alias{wbt_create_rectangular_vector_grid}
\title{Create rectangular vector grid}
\usage{
wbt_create_rectangular_vector_grid(
  input,
  output,
  width,
  height,
  xorig = 0,
  yorig = 0,
  wd = NULL,
  verbose_mode = NULL,
  compress_rasters = NULL,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input base file.}

\item{output}{Output vector polygon file.}

\item{width}{The grid cell width.}

\item{height}{The grid cell height.}

\item{xorig}{The grid origin x-coordinate.}

\item{yorig}{The grid origin y-coordinate.}

\item{wd}{Changes the working directory. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_wd()} for details.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_verbose()} for details.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_compress_rasters()} for details.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool. Default: \code{FALSE}.}
}
\value{
Returns the tool text outputs.
}
\description{
Creates a rectangular vector grid.
}
\keyword{GISAnalysis}
