% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{wbt_cross_tabulation}
\alias{wbt_cross_tabulation}
\title{Cross tabulation}
\usage{
wbt_cross_tabulation(input1, input2, output, wd = NULL, verbose_mode = FALSE)
}
\arguments{
\item{input1}{Input raster file 1.}

\item{input2}{Input raster file 1.}

\item{output}{Output HTML file (default name will be based on input file if unspecified).}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a cross-tabulation on two categorical images.
}
