% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{wbt_csv_points_to_vector}
\alias{wbt_csv_points_to_vector}
\title{Csv points to vector}
\usage{
wbt_csv_points_to_vector(
  input,
  output,
  xfield = 0,
  yfield = 1,
  epsg = NULL,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input CSV file (i.e. source of data to be imported).}

\item{output}{Output vector file.}

\item{xfield}{X field number (e.g. 0 for first field).}

\item{yfield}{Y field number (e.g. 1 for second field).}

\item{epsg}{EPSG projection (e.g. 2958).}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Converts a CSV text file to vector points.
}
