% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{wbt_cumulative_distribution}
\alias{wbt_cumulative_distribution}
\title{Cumulative distribution}
\usage{
wbt_cumulative_distribution(input, output, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Converts a raster image to its cumulative distribution function.
}
\examples{
\dontrun{
dem <- system.file("extdata", "DEM.tif", package="whitebox")
wbt_cumulative_distribution(input = dem, output = 'output.tif')
}
}
