% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{wbt_d8_mass_flux}
\alias{wbt_d8_mass_flux}
\title{D8 mass flux}
\usage{
wbt_d8_mass_flux(dem, loading, efficiency, absorption, output,
  verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{loading}{Input loading raster file.}

\item{efficiency}{Input efficiency raster file.}

\item{absorption}{Input absorption raster file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a D8 mass flux calculation.
}
