% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{wbt_d_inf_pointer}
\alias{wbt_d_inf_pointer}
\title{D inf pointer}
\usage{
wbt_d_inf_pointer(dem, output, wd = NULL, verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates a D-infinity flow pointer (flow direction) raster from an input DEM.
}
