% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{wbt_downslope_distance_to_stream}
\alias{wbt_downslope_distance_to_stream}
\title{Downslope distance to stream}
\usage{
wbt_downslope_distance_to_stream(
  dem,
  streams,
  output,
  dinf = FALSE,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{streams}{Input raster streams file.}

\item{output}{Output raster file.}

\item{dinf}{Use the D-infinity flow algoirthm instead of D8?.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
Measures distance to the nearest downslope stream cell.
}
