% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{wbt_downslope_flowpath_length}
\alias{wbt_downslope_flowpath_length}
\title{Downslope flowpath length}
\usage{
wbt_downslope_flowpath_length(d8_pntr, output, watersheds = NULL,
  weights = NULL, esri_pntr = FALSE, verbose_mode = FALSE)
}
\arguments{
\item{d8_pntr}{Input D8 pointer raster file.}

\item{output}{Output raster file.}

\item{watersheds}{Optional input watershed raster file.}

\item{weights}{Optional input weights raster file.}

\item{esri_pntr}{D8 pointer uses the ESRI style scheme.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the downslope flowpath length from each cell to basin outlet.
}
