% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis_filters.R
\name{wbt_edge_preserving_mean_filter}
\alias{wbt_edge_preserving_mean_filter}
\title{Edge preserving mean filter
Performs a simple edge-preserving mean filter on an input image.}
\usage{
wbt_edge_preserving_mean_filter(
  input,
  output,
  threshold,
  filter = 11,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{threshold}{Maximum difference in values.}

\item{filter}{Size of the filter kernel.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Edge preserving mean filter
Performs a simple edge-preserving mean filter on an input image.
}
\keyword{ImageProcessingToolsFilters}
