% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_elev_relative_to_watershed_min_max}
\alias{wbt_elev_relative_to_watershed_min_max}
\title{Elev relative to watershed min max}
\usage{
wbt_elev_relative_to_watershed_min_max(dem, watersheds, output,
  verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{watersheds}{Input raster watersheds file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the elevation of a location relative to the minimum and maximum elevations in a watershed.
}
