% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_erase_polygon_from_lidar}
\alias{wbt_erase_polygon_from_lidar}
\title{Erase polygon from lidar}
\usage{
wbt_erase_polygon_from_lidar(input, polygons, output, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input LiDAR file.}

\item{polygons}{Input vector polygons file.}

\item{output}{Output LiDAR file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Erases (cuts out) a vector polygon or polygons from a LiDAR point cloud.
}
