% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_euclidean_distance}
\alias{wbt_euclidean_distance}
\title{Euclidean distance}
\usage{
wbt_euclidean_distance(input, output, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the Shih and Wu (2004) Euclidean distance transform.
}
