% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbt.R
\name{wbt_exe_path}
\alias{wbt_exe_path}
\title{File path of the WhiteboxTools executable.}
\usage{
wbt_exe_path()
}
\value{
Returns the file path of WhiteboxTools executable.
}
\description{
Get the file path of the WhiteboxTools executable.
}
\examples{
wbt_exe <- wbt_exe_path()
}
