% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{wbt_export_table_to_csv}
\alias{wbt_export_table_to_csv}
\title{Export table to csv}
\usage{
wbt_export_table_to_csv(
  input,
  output,
  headers = TRUE,
  wd = NULL,
  verbose_mode = FALSE
)
}
\arguments{
\item{input}{Input vector file.}

\item{output}{Output raster file.}

\item{headers}{Export field names as file header?.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Exports an attribute table to a CSV text file.
}
