% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_fetch_analysis}
\alias{wbt_fetch_analysis}
\title{Fetch analysis
Performs an analysis of fetch or upwind distance to an obstacle.}
\usage{
wbt_fetch_analysis(
  dem,
  output,
  azimuth = 0,
  hgt_inc = 0.05,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{azimuth}{Wind azimuth in degrees in degrees.}

\item{hgt_inc}{Height increment value.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Fetch analysis
Performs an analysis of fetch or upwind distance to an obstacle.
}
\keyword{GeomorphometricAnalysis}
