% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{wbt_fill_burn}
\alias{wbt_fill_burn}
\title{Fill burn
Burns streams into a DEM using the FillBurn (Saunders, 1999) method.}
\usage{
wbt_fill_burn(
  dem,
  streams,
  output,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{streams}{Input vector streams file.}

\item{output}{Output raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Fill burn
Burns streams into a DEM using the FillBurn (Saunders, 1999) method.
}
\keyword{HydrologicalAnalysis}
