% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{wbt_fill_depressions_wang_and_liu}
\alias{wbt_fill_depressions_wang_and_liu}
\title{Fill depressions wang and liu}
\usage{
wbt_fill_depressions_wang_and_liu(
  dem,
  output,
  fix_flats = TRUE,
  flat_increment = NULL,
  wd = NULL,
  verbose_mode = FALSE
)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{fix_flats}{Optional flag indicating whether flat areas should have a small gradient applied.}

\item{flat_increment}{Optional elevation increment applied to flat areas.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Fills all of the depressions in a DEM using the Wang and Liu (2006) method. Depression breaching should be preferred in most cases.
}
