% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_find_flightline_edge_points}
\alias{wbt_find_flightline_edge_points}
\title{Find flightline edge points}
\usage{
wbt_find_flightline_edge_points(input, output, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input LiDAR file.}

\item{output}{Output file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Identifies points along a flightline's edge in a LAS file.
}
