% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{wbt_find_parallel_flow}
\alias{wbt_find_parallel_flow}
\title{Find parallel flow}
\usage{
wbt_find_parallel_flow(d8_pntr, streams, output, verbose_mode = FALSE)
}
\arguments{
\item{d8_pntr}{Input D8 pointer raster file.}

\item{streams}{Input raster streams file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Finds areas of parallel flow in D8 flow direction rasters.
}
