% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{wbt_flatten_lakes}
\alias{wbt_flatten_lakes}
\title{Flatten lakes}
\usage{
wbt_flatten_lakes(dem, lakes, output, wd = NULL, verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{lakes}{Input lakes vector polygons file.}

\item{output}{Output raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Flattens lake polygons in a raster DEM.
}
