% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis_classification.R
\name{wbt_generalize_with_similarity}
\alias{wbt_generalize_with_similarity}
\title{Generalize with similarity}
\usage{
wbt_generalize_with_similarity(
  input,
  similarity,
  output,
  min_size = 4,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Name of the input raster image file.}

\item{similarity}{Names of the input similarity images.}

\item{output}{Name of the output raster file.}

\item{min_size}{Minimum feature size, in grid cells.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Generalizes a raster containing class or object features by removing small features using similarity criteria of neighbouring features.
}
