% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{wbt_histogram_matching_two_images}
\alias{wbt_histogram_matching_two_images}
\title{Histogram matching two images}
\usage{
wbt_histogram_matching_two_images(input1, input2, output, verbose_mode = FALSE)
}
\arguments{
\item{input1}{Input raster file to modify.}

\item{input2}{Input reference raster file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
This tool alters the cumulative distribution function of a raster image to that of another image.
}
