% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{wbt_in_place_add}
\alias{wbt_in_place_add}
\title{In place add}
\usage{
wbt_in_place_add(input1, input2, verbose_mode = FALSE)
}
\arguments{
\item{input1}{Input raster file.}

\item{input2}{Input raster file or constant value.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs an in-place addition operation (input1 += input2).
}
