% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{wbt_in_place_divide}
\alias{wbt_in_place_divide}
\title{In place divide}
\usage{
wbt_in_place_divide(input1, input2, wd = NULL, verbose_mode = FALSE)
}
\arguments{
\item{input1}{Input raster file.}

\item{input2}{Input raster file or constant value.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs an in-place division operation (input1 /= input2).
}
