% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{wbt_in_place_subtract}
\alias{wbt_in_place_subtract}
\title{In place subtract
Performs an in-place subtraction operation (input1 -= input2).}
\usage{
wbt_in_place_subtract(
  input1,
  input2,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input1}{Input raster file.}

\item{input2}{Input raster file or constant value.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
In place subtract
Performs an in-place subtraction operation (input1 -= input2).
}
\keyword{MathandStatsTools}
