% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{wbt_insert_dams}
\alias{wbt_insert_dams}
\title{Insert dams}
\usage{
wbt_insert_dams(dem, dam_pts, output, damlength, verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{dam_pts}{Input vector dam points file.}

\item{output}{Output file.}

\item{damlength}{Maximum length of the dam.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the impoundment size resulting from damming a DEM.
}
