% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbt.R
\name{wbt_install}
\alias{wbt_install}
\alias{install_whitebox}
\title{Download and Install WhiteboxTools}
\usage{
wbt_install(pkg_dir = find.package("whitebox"))
}
\arguments{
\item{pkg_dir}{default install path is to whitebox package "WBT" folder}
}
\value{
Prints out the location of the WhiteboxTools binary
}
\description{
This function downloads the WhiteboxTools binary if needed.
}
\examples{
\dontrun{
wbt_install()
}
}
