% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{wbt_integer_division}
\alias{wbt_integer_division}
\title{Integer division}
\usage{
wbt_integer_division(input1, input2, output, wd = NULL, verbose_mode = FALSE)
}
\arguments{
\item{input1}{Input raster file or constant value.}

\item{input2}{Input raster file or constant value.}

\item{output}{Output raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs an integer division operation on two rasters or a raster and a constant value.
}
