% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{wbt_inverse_principal_component_analysis}
\alias{wbt_inverse_principal_component_analysis}
\title{Inverse principal component analysis}
\usage{
wbt_inverse_principal_component_analysis(
  inputs,
  report,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{inputs}{Name of the input PCA component images.}

\item{report}{Name of the PCA report file (*.html).}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
This tool performs an inverse principal component analysis on a series of input component images.
}
