% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{wbt_knn_regression}
\alias{wbt_knn_regression}
\title{Knn regression}
\usage{
wbt_knn_regression(
  inputs,
  training,
  field,
  scaling = "Normalize",
  output = NULL,
  k = 5,
  weight = TRUE,
  test_proportion = 0.2,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{inputs}{Names of the input predictor rasters.}

\item{training}{Name of the input training site points Shapefile.}

\item{field}{Name of the attribute containing response variable name data.}

\item{scaling}{Scaling method for predictors. Options include 'None', 'Normalize', and 'Standardize'.}

\item{output}{Name of the output raster file.}

\item{k}{k-parameter, which determines the number of nearest neighbours used.}

\item{weight}{Use distance weighting?.}

\item{test_proportion}{The proportion of the dataset to include in the test split; default is 0.2.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a supervised k-nearest neighbour regression using training site points and predictor rasters.
}
