% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{wbt_ks_test_for_normality}
\alias{wbt_ks_test_for_normality}
\title{Ks test for normality
Evaluates whether the values in a raster are normally distributed.}
\usage{
wbt_ks_test_for_normality(
  input,
  output,
  num_samples = NULL,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output HTML file.}

\item{num_samples}{Number of samples. Leave blank to use whole image.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Ks test for normality
Evaluates whether the values in a raster are normally distributed.
}
\keyword{MathandStatsTools}
