% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_las_to_laz}
\alias{wbt_las_to_laz}
\title{Las to laz}
\usage{
wbt_las_to_laz(
  input,
  output = NULL,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{input}{Name of the input LAS files (leave blank to use all LAS files in WorkingDirectory.}

\item{output}{Output LAZ file (including extension).}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
This tool converts one or more LAS files into the LAZ format.
}
