% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_las_to_zlidar}
\alias{wbt_las_to_zlidar}
\title{Las to zlidar}
\usage{
wbt_las_to_zlidar(
  inputs = NULL,
  outdir = NULL,
  wd = NULL,
  verbose_mode = FALSE
)
}
\arguments{
\item{inputs}{Input LAS files.}

\item{outdir}{Output directory into which zlidar files are created. If unspecified, it is assumed to be the same as the inputs.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Converts one or more LAS files into the zlidar compressed LiDAR data format.
}
