% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbt.R
\name{wbt_launch_runner}
\alias{wbt_launch_runner}
\title{Launch 'WhiteboxTools Runner' GUI}
\usage{
wbt_launch_runner(clear_app_state = FALSE)
}
\arguments{
\item{clear_app_state}{Clear application state memory? Default: \code{FALSE}}
}
\description{
\code{wbt_launch_runner()}: Launch the 'WhiteboxTools Runner' GUI at \code{wbt_runner_path()}
}
\details{
Opens the 'WhiteboxTools Runner' GUI included with WhiteboxTools Open Core v2.3.0 or higher.
}
\seealso{
\code{\link[=wbt_runner_path]{wbt_runner_path()}}
}
\keyword{General}
