% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whitebox_utilities.R
\name{wbt_launch_wb_runner}
\alias{wbt_launch_wb_runner}
\title{Launch wb runner}
\usage{
wbt_launch_wb_runner(
  clear_app_state = FALSE,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{clear_app_state}{Clear the application state memory?.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Opens the Whitebox Runner application.
}
\keyword{WhiteboxUtilities}
