% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{wbt_less_than}
\alias{wbt_less_than}
\title{Less than
Performs a less-than comparison operation on two rasters or a raster and a constant value.}
\usage{
wbt_less_than(
  input1,
  input2,
  output,
  incl_equals = FALSE,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input1}{Input raster file or constant value.}

\item{input2}{Input raster file or constant value.}

\item{output}{Output raster file.}

\item{incl_equals}{Perform a less-than-or-equal-to operation.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Less than
Performs a less-than comparison operation on two rasters or a raster and a constant value.
}
\keyword{MathandStatsTools}
