% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_lidar_colourize}
\alias{wbt_lidar_colourize}
\title{Lidar colourize}
\usage{
wbt_lidar_colourize(in_lidar, in_image, output, verbose_mode = FALSE)
}
\arguments{
\item{in_lidar}{Input LiDAR file.}

\item{in_image}{Input colour image file.}

\item{output}{Output LiDAR file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Adds the red-green-blue colour fields of a LiDAR (LAS) file based on an input image.
}
