% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_lidar_ground_point_filter}
\alias{wbt_lidar_ground_point_filter}
\title{Lidar ground point filter}
\usage{
wbt_lidar_ground_point_filter(
  input,
  output,
  radius = 2,
  min_neighbours = 0,
  slope_threshold = 45,
  height_threshold = 1,
  classify = TRUE,
  slope_norm = TRUE,
  height_above_ground = FALSE,
  wd = NULL,
  verbose_mode = NULL,
  compress_rasters = NULL,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input LiDAR file.}

\item{output}{Output LiDAR file.}

\item{radius}{Search Radius.}

\item{min_neighbours}{The minimum number of neighbouring points within search areas. If fewer points than this threshold are identified during the fixed-radius search, a subsequent kNN search is performed to identify the k number of neighbours.}

\item{slope_threshold}{Maximum inter-point slope to be considered an off-terrain point.}

\item{height_threshold}{Inter-point height difference to be considered an off-terrain point.}

\item{classify}{Classify points as ground (2) or off-ground (1).}

\item{slope_norm}{Perform initial ground slope normalization?.}

\item{height_above_ground}{Transform output to height above average ground elevation?.}

\item{wd}{Changes the working directory. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_wd()} for details.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_verbose()} for details.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_compress_rasters()} for details.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool. Default: \code{FALSE}.}
}
\value{
Returns the tool text outputs.
}
\description{
Identifies ground points within LiDAR dataset using a slope-based method.
}
\keyword{LiDARTools}
