% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_lidar_hillshade}
\alias{wbt_lidar_hillshade}
\title{Lidar hillshade}
\usage{
wbt_lidar_hillshade(
  input,
  output,
  azimuth = 315,
  altitude = 30,
  radius = 1,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input LiDAR file.}

\item{output}{Output file.}

\item{azimuth}{Illumination source azimuth in degrees.}

\item{altitude}{Illumination source altitude in degrees.}

\item{radius}{Search Radius.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates a hillshade value for points within a LAS file and stores these data in the RGB field.
}
\keyword{LiDARTools}
