% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_lidar_thin}
\alias{wbt_lidar_thin}
\title{Lidar thin}
\usage{
wbt_lidar_thin(
  input,
  output,
  resolution = 2,
  method = "lowest",
  save_filtered = FALSE,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input LiDAR file.}

\item{output}{Output LiDAR file.}

\item{resolution}{The size of the square area used to evaluate nearby points in the LiDAR data.}

\item{method}{Point selection method; options are 'first', 'last', 'lowest' (default), 'highest', 'nearest'.}

\item{save_filtered}{Save filtered points to separate file?.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Thins a LiDAR point cloud, reducing point density.
}
\keyword{LiDARTools}
