% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_lidar_thin_high_density}
\alias{wbt_lidar_thin_high_density}
\title{Lidar thin high density}
\usage{
wbt_lidar_thin_high_density(
  input,
  output,
  density,
  resolution = 1,
  save_filtered = FALSE,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{input}{Input LiDAR file.}

\item{output}{Output LiDAR file.}

\item{density}{Max. point density (points / m^3).}

\item{resolution}{Output raster's grid resolution.}

\item{save_filtered}{Save filtered points to separate file?.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
Thins points from high density areas within a LiDAR point cloud.
}
