% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_network_analysis.R
\name{wbt_long_profile_from_points}
\alias{wbt_long_profile_from_points}
\title{Long profile from points}
\usage{
wbt_long_profile_from_points(
  d8_pntr,
  points,
  dem,
  output,
  esri_pntr = FALSE,
  verbose_mode = FALSE
)
}
\arguments{
\item{d8_pntr}{Input raster D8 pointer file.}

\item{points}{Input vector points file.}

\item{dem}{Input raster DEM file.}

\item{output}{Output HTML file.}

\item{esri_pntr}{D8 pointer uses the ESRI style scheme.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Plots the longitudinal profiles from flow-paths initiating from a set of vector points.
}
