% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{wbt_low_points_on_headwater_divides}
\alias{wbt_low_points_on_headwater_divides}
\title{Low points on headwater divides}
\usage{
wbt_low_points_on_headwater_divides(
  dem,
  streams,
  output,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{dem}{Name of the input DEM raster file.}

\item{streams}{Name of the input stream channel raster file.}

\item{output}{Name of the output vector file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
This tool locates saddle points along ridges within a digital elevation model (DEM).
}
