% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_max_absolute_overlay}
\alias{wbt_max_absolute_overlay}
\title{Max absolute overlay}
\usage{
wbt_max_absolute_overlay(inputs, output, verbose_mode = FALSE)
}
\arguments{
\item{inputs}{Input raster files.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Evaluates the maximum absolute value for each grid cell from a stack of input rasters.
}
