% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_max_anisotropy_dev}
\alias{wbt_max_anisotropy_dev}
\title{Max anisotropy dev}
\usage{
wbt_max_anisotropy_dev(
  dem,
  out_mag,
  out_scale,
  max_scale,
  min_scale = 3,
  step = 2,
  verbose_mode = FALSE
)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{out_mag}{Output raster DEVmax magnitude file.}

\item{out_scale}{Output raster DEVmax scale file.}

\item{max_scale}{Maximum search neighbourhood radius in grid cells.}

\item{min_scale}{Minimum search neighbourhood radius in grid cells.}

\item{step}{Step size as any positive non-zero integer.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the maximum anisotropy (directionality) in elevation deviation over a range of spatial scales.
}
