% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_max_anisotropy_dev_signature}
\alias{wbt_max_anisotropy_dev_signature}
\title{Max anisotropy dev signature}
\usage{
wbt_max_anisotropy_dev_signature(dem, points, output, max_scale,
  min_scale = 1, step = 1, verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{points}{Input vector points file.}

\item{output}{Output HTML file.}

\item{max_scale}{Maximum search neighbourhood radius in grid cells.}

\item{min_scale}{Minimum search neighbourhood radius in grid cells.}

\item{step}{Step size as any positive non-zero integer.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the anisotropy in deviation from mean for points over a range of spatial scales.
}
