% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{wbt_max_upslope_value}
\alias{wbt_max_upslope_value}
\title{Max upslope value}
\usage{
wbt_max_upslope_value(
  dem,
  values,
  output,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{dem}{Input DEM; it must be depressionless.}

\item{values}{Name of the input values raster file.}

\item{output}{Name of the output raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the maximum upslope value from an input values raster along flowpaths.
}
\keyword{HydrologicalAnalysis}
