% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{wbt_merge_vectors}
\alias{wbt_merge_vectors}
\title{Merge vectors}
\usage{
wbt_merge_vectors(inputs, output, verbose_mode = FALSE)
}
\arguments{
\item{inputs}{Input vector files.}

\item{output}{Output vector file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Combines two or more input vectors of the same ShapeType creating a single, new output vector.
}
