% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis_classification.R
\name{wbt_min_dist_classification}
\alias{wbt_min_dist_classification}
\title{Min dist classification}
\usage{
wbt_min_dist_classification(
  inputs,
  polys,
  field,
  output,
  threshold = NULL,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{inputs}{Names of the input band images.}

\item{polys}{Name of the input training site polygons shapefile.}

\item{field}{Name of the attribute containing class name data.}

\item{output}{Name of the output raster file.}

\item{threshold}{Distance threshold, in z-scores; blank for none.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a supervised minimum-distance classification using training site polygons and multi-spectral images.
}
