% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_minimum_bounding_circle}
\alias{wbt_minimum_bounding_circle}
\title{Minimum bounding circle}
\usage{
wbt_minimum_bounding_circle(
  input,
  output,
  features = TRUE,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input vector file.}

\item{output}{Output vector polygon file.}

\item{features}{Find the minimum bounding circle around each individual vector feature.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Delineates the minimum bounding circle (i.e. smallest enclosing circle) for a group of vectors.
}
\keyword{GISAnalysis}
