% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{wbt_modify_no_data_value}
\alias{wbt_modify_no_data_value}
\title{Modify no data value}
\usage{
wbt_modify_no_data_value(input, new_value = "-32768.0",
  verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{new_value}{New NoData value.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Converts nodata values in a raster to zero.
}
