% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{wbt_mosaic_with_feathering}
\alias{wbt_mosaic_with_feathering}
\title{Mosaic with feathering
Mosaics two images together using a feathering technique in overlapping areas to reduce edge-effects.}
\usage{
wbt_mosaic_with_feathering(
  input1,
  input2,
  output,
  method = "cc",
  weight = 4,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input1}{Input raster file to modify.}

\item{input2}{Input reference raster file.}

\item{output}{Output raster file.}

\item{method}{Resampling method; options include 'nn' (nearest neighbour), 'bilinear', and 'cc' (cubic convolution).}

\item{weight}{.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Mosaic with feathering
Mosaics two images together using a feathering technique in overlapping areas to reduce edge-effects.
}
\keyword{ImageProcessingTools}
