% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{wbt_multi_part_to_single_part}
\alias{wbt_multi_part_to_single_part}
\title{Multi part to single part}
\usage{
wbt_multi_part_to_single_part(
  input,
  output,
  exclude_holes = TRUE,
  wd = NULL,
  verbose_mode = NULL,
  compress_rasters = NULL,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input vector line or polygon file.}

\item{output}{Output vector line or polygon file.}

\item{exclude_holes}{Exclude hole parts from the feature splitting? (holes will continue to belong to their features in output.).}

\item{wd}{Changes the working directory. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_wd()} for details.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_verbose()} for details.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_compress_rasters()} for details.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool. Default: \code{FALSE}.}
}
\value{
Returns the tool text outputs.
}
\description{
Converts a vector file containing multi-part features into a vector containing only single-part features.
}
\keyword{DataTools}
