% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_multidirectional_hillshade}
\alias{wbt_multidirectional_hillshade}
\title{Multidirectional hillshade}
\usage{
wbt_multidirectional_hillshade(
  dem,
  output,
  altitude = 45,
  zfactor = 1,
  full_mode = FALSE,
  wd = NULL,
  verbose_mode = FALSE
)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{altitude}{Illumination source altitude in degrees.}

\item{zfactor}{Optional multiplier for when the vertical and horizontal units are not the same.}

\item{full_mode}{Optional flag indicating whether to use full 360-degrees of illumination sources.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates a multi-direction hillshade raster from an input DEM.
}
