% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_normalize_lidar}
\alias{wbt_normalize_lidar}
\title{Normalize lidar}
\usage{
wbt_normalize_lidar(
  input,
  output,
  dtm,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Name of the input LiDAR file.}

\item{output}{Name of the output LiDAR file.}

\item{dtm}{Name of the input digital terrain model (DTM) raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Normalizes a LiDAR point cloud.
}
\keyword{LiDARTools}
