% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_num_downslope_neighbours}
\alias{wbt_num_downslope_neighbours}
\title{Num downslope neighbours
Calculates the number of downslope neighbours to each grid cell in a DEM.}
\usage{
wbt_num_downslope_neighbours(
  dem,
  output,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Num downslope neighbours
Calculates the number of downslope neighbours to each grid cell in a DEM.
}
\keyword{GeomorphometricAnalysis}
