% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{wbt_parallelepiped_classification}
\alias{wbt_parallelepiped_classification}
\title{Parallelepiped classification}
\usage{
wbt_parallelepiped_classification(
  inputs,
  polys,
  field,
  output,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{inputs}{Name of the input band images.}

\item{polys}{Name of the input training site polygons shapefile.}

\item{field}{Name of the attribute containing class name data.}

\item{output}{Name of the output raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a supervised parallelepiped classification using training site polygons and multi-spectral images.
}
