% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_pennock_landform_class}
\alias{wbt_pennock_landform_class}
\title{Pennock landform class
Classifies hillslope zones based on slope, profile curvature, and plan curvature.}
\usage{
wbt_pennock_landform_class(
  dem,
  output,
  slope = 3,
  prof = 0.1,
  plan = 0,
  zfactor = NULL,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{slope}{Slope threshold value, in degrees (default is 3.0).}

\item{prof}{Profile curvature threshold value (default is 0.1).}

\item{plan}{Plan curvature threshold value (default is 0.0).}

\item{zfactor}{Optional multiplier for when the vertical and horizontal units are not the same.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Pennock landform class
Classifies hillslope zones based on slope, profile curvature, and plan curvature.
}
\keyword{GeomorphometricAnalysis}
