% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_percent_greater_than}
\alias{wbt_percent_greater_than}
\title{Percent greater than}
\usage{
wbt_percent_greater_than(inputs, comparison, output, verbose_mode = FALSE)
}
\arguments{
\item{inputs}{Input raster files.}

\item{comparison}{Input comparison raster file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the percentage of a raster stack that have cell values greather than an input on a cell-by-cell basis.
}
