% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_perimeter_area_ratio}
\alias{wbt_perimeter_area_ratio}
\title{Perimeter area ratio}
\usage{
wbt_perimeter_area_ratio(input, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input vector polygon file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the perimeter-area ratio of vector polygons.
}
