% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_pick_from_list}
\alias{wbt_pick_from_list}
\title{Pick from list}
\usage{
wbt_pick_from_list(inputs, pos_input, output, wd = NULL, verbose_mode = FALSE)
}
\arguments{
\item{inputs}{Input raster files.}

\item{pos_input}{Input position raster file.}

\item{output}{Output raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Outputs the value from a raster stack specified by a position raster.
}
