% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_polygon_area}
\alias{wbt_polygon_area}
\title{Polygon area}
\usage{
wbt_polygon_area(input, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input vector polygon file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the area of vector polygons.
}
