% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_polygon_short_axis}
\alias{wbt_polygon_short_axis}
\title{Polygon short axis}
\usage{
wbt_polygon_short_axis(input, output, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input vector polygons file.}

\item{output}{Output vector polygon file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
This tool can be used to map the short axis of polygon features.
}
