% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_profile}
\alias{wbt_profile}
\title{Profile}
\usage{
wbt_profile(
  lines,
  surface,
  output,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{lines}{Input vector line file.}

\item{surface}{Input raster surface file.}

\item{output}{Output HTML file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
Plots profiles from digital surface models.
}
