% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{wbt_qin_flow_accumulation}
\alias{wbt_qin_flow_accumulation}
\title{Qin flow accumulation}
\usage{
wbt_qin_flow_accumulation(
  dem,
  output,
  out_type = "specific contributing area",
  exponent = 10,
  max_slope = 45,
  threshold = NULL,
  log = FALSE,
  clip = FALSE,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{dem}{Name of the input DEM raster file; must be depressionless.}

\item{output}{Name of the output upslope saturated area file.}

\item{out_type}{Output type; one of 'cells', 'specific contributing area' (default), and 'catchment area'.}

\item{exponent}{Optional upper-bound exponent parameter; default is 10.0.}

\item{max_slope}{Optional upper-bound slope parameter, in degrees (0-90); default is 45.0.}

\item{threshold}{Optional convergence threshold parameter, in grid cells; default is inifinity.}

\item{log}{Log-transform the output values?.}

\item{clip}{Optional flag to request clipping the display max by 1 percent.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
This tool calculates Qin et al. (2007) flow accumulation.
}
