% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{wbt_random_sample}
\alias{wbt_random_sample}
\title{Random sample
Creates an image containing randomly located sample grid cells with unique IDs.}
\usage{
wbt_random_sample(
  base,
  output,
  num_samples = 1000,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{base}{Input raster file.}

\item{output}{Output raster file.}

\item{num_samples}{Number of samples.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Random sample
Creates an image containing randomly located sample grid cells with unique IDs.
}
\keyword{MathandStatsTools}
