% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_raster_area}
\alias{wbt_raster_area}
\title{Raster area}
\usage{
wbt_raster_area(input, output = NULL, out_text = FALSE,
  units = "grid cells", zero_back = FALSE, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{out_text}{Would you like to output polygon areas to text?.}

\item{units}{Area units; options include 'grid cells' and 'map units'.}

\item{zero_back}{Flag indicating whether zero values should be treated as a background.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the area of polygons or classes within a raster image.
}
