% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_raster_perimeter}
\alias{wbt_raster_perimeter}
\title{Raster perimeter}
\usage{
wbt_raster_perimeter(
  input,
  output = NULL,
  out_text = FALSE,
  units = "grid cells",
  zero_back = FALSE,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{out_text}{Would you like to output polygon areas to text?.}

\item{units}{Area units; options include 'grid cells' and 'map units'.}

\item{zero_back}{Flag indicating whether zero values should be treated as a background.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the perimeters of polygons or classes within a raster image.
}
