% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{wbt_raster_summary_stats}
\alias{wbt_raster_summary_stats}
\title{Raster summary stats}
\usage{
wbt_raster_summary_stats(input, wd = NULL, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Measures a rasters min, max, average, standard deviation, num. non-nodata cells, and total.
}
