% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_network_analysis.R
\name{wbt_rasterize_streams}
\alias{wbt_rasterize_streams}
\title{Rasterize streams}
\usage{
wbt_rasterize_streams(
  streams,
  base,
  output,
  nodata = TRUE,
  feature_id = FALSE,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{streams}{Input vector streams file.}

\item{base}{Input base raster file.}

\item{output}{Output raster file.}

\item{nodata}{Use NoData value for background?.}

\item{feature_id}{Use feature number as output value?.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Rasterizes vector streams based on Lindsay (2016) method.
}
\keyword{StreamNetworkAnalysis}
