% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_reclass_equal_interval}
\alias{wbt_reclass_equal_interval}
\title{Reclass equal interval
Reclassifies the values in a raster image based on equal-ranges.}
\usage{
wbt_reclass_equal_interval(
  input,
  output,
  interval = 10,
  start_val = NULL,
  end_val = NULL,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{interval}{Class interval size.}

\item{start_val}{Optional starting value (default is input minimum value).}

\item{end_val}{Optional ending value (default is input maximum value).}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Reclass equal interval
Reclassifies the values in a raster image based on equal-ranges.
}
\keyword{GISAnalysis}
