% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_agriculture.R
\name{wbt_recreate_pass_lines}
\alias{wbt_recreate_pass_lines}
\title{Recreate pass lines
This tool can be used to approximate the harvester pass lines from yield points.}
\usage{
wbt_recreate_pass_lines(
  input,
  yield_field_name,
  output_lines,
  output_points,
  max_change_in_heading = 25,
  ignore_zeros = FALSE,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Name of the input points shapefile.}

\item{yield_field_name}{Name of the attribute containing yield data.}

\item{output_lines}{Name of the output pass lines shapefile.}

\item{output_points}{Name of the output points shapefile.}

\item{max_change_in_heading}{Max change in heading.}

\item{ignore_zeros}{Ignore zero-valued yield points?.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Recreate pass lines
This tool can be used to approximate the harvester pass lines from yield points.
}
\keyword{PrecisionAgriculture}
