% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_remove_off_terrain_objects}
\alias{wbt_remove_off_terrain_objects}
\title{Remove off terrain objects
Removes off-terrain objects from a raster digital elevation model (DEM).}
\usage{
wbt_remove_off_terrain_objects(
  dem,
  output,
  filter = 11,
  slope = 15,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{filter}{Filter size (cells).}

\item{slope}{Slope threshold value.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Remove off terrain objects
Removes off-terrain objects from a raster digital elevation model (DEM).
}
\keyword{GeomorphometricAnalysis}
