% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{wbt_remove_polygon_holes}
\alias{wbt_remove_polygon_holes}
\title{Remove polygon holes}
\usage{
wbt_remove_polygon_holes(input, output, wd = NULL, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input vector polygon file.}

\item{output}{Output vector polygon file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Removes holes within the features of a vector polygon file.
}
