% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{wbt_rgb_to_ihs}
\alias{wbt_rgb_to_ihs}
\title{Rgb to ihs
Converts red, green, and blue (RGB) images into intensity, hue, and saturation (IHS) images.}
\usage{
wbt_rgb_to_ihs(
  intensity,
  hue,
  saturation,
  red = NULL,
  green = NULL,
  blue = NULL,
  composite = NULL,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{intensity}{Output intensity raster file.}

\item{hue}{Output hue raster file.}

\item{saturation}{Output saturation raster file.}

\item{red}{Input red band image file. Optionally specified if colour-composite not specified.}

\item{green}{Input green band image file. Optionally specified if colour-composite not specified.}

\item{blue}{Input blue band image file. Optionally specified if colour-composite not specified.}

\item{composite}{Input colour-composite image file. Only used if individual bands are not specified.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Rgb to ihs
Converts red, green, and blue (RGB) images into intensity, hue, and saturation (IHS) images.
}
\keyword{ImageProcessingTools}
