% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbt.R
\name{wbt_rust_backtrace}
\alias{wbt_rust_backtrace}
\title{Convenience method for setting RUST_BACKTRACE options for debugging}
\usage{
wbt_rust_backtrace(RUST_BACKTRACE = c("0", "1", "full"))
}
\arguments{
\item{RUST_BACKTRACE}{One of \code{"0"}, \code{"1"}, \code{"full"}, Logical values are converted to integer and then character.}
}
\value{
value of system environment variable \code{RUST_BACKTRACE}
}
\description{
Convenience method for setting RUST_BACKTRACE options for debugging
}
\examples{
\dontrun{
wbt_rust_backtrace(TRUE)
}
}
\keyword{General}
