% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_select_tiles_by_polygon}
\alias{wbt_select_tiles_by_polygon}
\title{Select tiles by polygon}
\usage{
wbt_select_tiles_by_polygon(
  indir,
  outdir,
  polygons,
  wd = NULL,
  verbose_mode = NULL,
  compress_rasters = NULL,
  command_only = FALSE
)
}
\arguments{
\item{indir}{Input LAS file source directory.}

\item{outdir}{Output directory into which LAS files within the polygon are copied.}

\item{polygons}{Input vector polygons file.}

\item{wd}{Changes the working directory. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_wd()} for details.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_verbose()} for details.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_compress_rasters()} for details.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool. Default: \code{FALSE}.}
}
\value{
Returns the tool text outputs.
}
\description{
Copies LiDAR tiles overlapping with a polygon into an output directory.
}
\keyword{LiDARTools}
