% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_shadow_image}
\alias{wbt_shadow_image}
\title{Shadow image}
\usage{
wbt_shadow_image(
  input,
  output,
  palette = "soft",
  max_dist = "",
  date = "21/06/2021",
  time = "13:00",
  location = "43.5448/-80.2482/-4",
  wd = NULL,
  verbose_mode = NULL,
  compress_rasters = NULL,
  command_only = FALSE
)
}
\arguments{
\item{input}{Name of the input digital surface model (DSM) raster file.}

\item{output}{Name of the output raster file.}

\item{palette}{DSM image palette; options are 'atlas', 'high_relief', 'arid', 'soft', 'muted', 'light_quant', 'purple', 'viridi', 'gn_yl', 'pi_y_g', 'bl_yl_rd', 'deep', and 'none'.}

\item{max_dist}{Optional maximum search distance, in xy unites. Minimum value is 5 x cell size.}

\item{date}{Date in format DD/MM/YYYY.}

\item{time}{Time in format HH::MM, e.g. 03:15AM or 14:30.}

\item{location}{Location, defined as Lat/Long/UTC-offset (e.g. 43.5448/-80.2482/-4).}

\item{wd}{Changes the working directory. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_wd()} for details.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_verbose()} for details.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_compress_rasters()} for details.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool. Default: \code{FALSE}.}
}
\value{
Returns the tool text outputs.
}
\description{
This tool creates a raster of shadow areas based on an input DEM.
}
\keyword{GeomorphometricAnalysis}
