% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_shape_complexity_index_raster}
\alias{wbt_shape_complexity_index_raster}
\title{Shape complexity index raster}
\usage{
wbt_shape_complexity_index_raster(input, output, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the complexity of raster polygons or classes.
}
