% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_network_analysis.R
\name{wbt_shreve_stream_magnitude}
\alias{wbt_shreve_stream_magnitude}
\title{Shreve stream magnitude}
\usage{
wbt_shreve_stream_magnitude(d8_pntr, streams, output, esri_pntr = FALSE,
  zero_background = FALSE, verbose_mode = FALSE)
}
\arguments{
\item{d8_pntr}{Input raster D8 pointer file.}

\item{streams}{Input raster streams file.}

\item{output}{Output raster file.}

\item{esri_pntr}{D8 pointer uses the ESRI style scheme.}

\item{zero_background}{Flag indicating whether a background value of zero should be used.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Assigns the Shreve stream magnitude to each link in a stream network.
}
