% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_slope_vs_elevation_plot}
\alias{wbt_slope_vs_elevation_plot}
\title{Slope vs elevation plot}
\usage{
wbt_slope_vs_elevation_plot(inputs, output, watershed = NULL,
  verbose_mode = FALSE)
}
\arguments{
\item{inputs}{Input DEM files.}

\item{output}{Output HTML file (default name will be based on input file if unspecified).}

\item{watershed}{Input watershed files (optional).}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Creates a slope vs. elevation plot for one or more DEMs.
}
