% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis_enhancement.R
\name{wbt_standard_deviation_contrast_stretch}
\alias{wbt_standard_deviation_contrast_stretch}
\title{Standard deviation contrast stretch}
\usage{
wbt_standard_deviation_contrast_stretch(
  input,
  output,
  stdev = 2,
  num_tones = 256,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{stdev}{Standard deviation clip value.}

\item{num_tones}{Number of tones in the output image.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a standard-deviation contrast stretch on input images.
}
