% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_standard_deviation_of_slope}
\alias{wbt_standard_deviation_of_slope}
\title{Standard deviation of slope}
\usage{
wbt_standard_deviation_of_slope(
  input,
  output,
  zfactor = NULL,
  filterx = 11,
  filtery = 11,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input raster DEM file.}

\item{output}{Output raster DEM file.}

\item{zfactor}{Optional multiplier for when the vertical and horizontal units are not the same.}

\item{filterx}{Size of the filter kernel in the x-direction.}

\item{filtery}{Size of the filter kernel in the y-direction.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the standard deviation of slope from an input DEM.
}
\keyword{GeomorphometricAnalysis}
