% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_surface_area_ratio}
\alias{wbt_surface_area_ratio}
\title{Surface area ratio}
\usage{
wbt_surface_area_ratio(dem, output, wd = NULL, verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates a the surface area ratio of each grid cell in an input DEM.
}
