% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_symmetrical_difference}
\alias{wbt_symmetrical_difference}
\title{Symmetrical difference}
\usage{
wbt_symmetrical_difference(
  input,
  overlay,
  output,
  snap = 0,
  verbose_mode = FALSE
)
}
\arguments{
\item{input}{Input vector file.}

\item{overlay}{Input overlay vector file.}

\item{output}{Output vector file.}

\item{snap}{Snap tolerance.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Outputs the features that occur in one of the two vector inputs but not both, i.e. no overlapping features.
}
