% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{wbt_thicken_raster_line}
\alias{wbt_thicken_raster_line}
\title{Thicken raster line}
\usage{
wbt_thicken_raster_line(input, output, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Thickens single-cell wide lines within a raster image.
}
