% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_tin_gridding}
\alias{wbt_tin_gridding}
\title{Tin gridding}
\usage{
wbt_tin_gridding(input, output, resolution, field = NULL,
  use_z = FALSE, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input vector points file.}

\item{output}{Output raster file.}

\item{resolution}{Output raster's grid resolution.}

\item{field}{Input field name in attribute table.}

\item{use_z}{Use the 'z' dimension of the Shapefile's geometry instead of an attribute field?.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Creates a raster grid based on a triangular irregular network (TIN) fitted to vector points.
}
