% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{wbt_to_radians}
\alias{wbt_to_radians}
\title{To radians}
\usage{
wbt_to_radians(input, output, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Converts a raster from degrees to radians.
}
\examples{
\dontrun{
dem <- system.file("extdata", "DEM.tif", package="whitebox")
wbt_to_radians(input = dem, output = 'output.tif')
}
}
