% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbt.R
\name{wbt_tool_parameters}
\alias{wbt_tool_parameters}
\title{Tool parameter descriptions for a specific tool.}
\usage{
wbt_tool_parameters(tool_name)
}
\arguments{
\item{tool_name}{The name of the tool.}
}
\value{
Returns the tool parameter descriptions for a specific tool.
}
\description{
Retrieves the tool parameter descriptions for a specific tool.
}
\examples{
wbt_tool_parameters("lidar_info")
}
