% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbt.R
\name{wbt_toolbox}
\alias{wbt_toolbox}
\title{The toolbox for a specific tool.}
\usage{
wbt_toolbox(tool_name = NULL)
}
\arguments{
\item{tool_name}{The name of the tool.}
}
\value{
Returns the toolbox for a specific tool.
}
\description{
Retrieve the toolbox for a specific tool.
}
\examples{
wbt_toolbox("breach_depressions")
}
