% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_travelling_salesman_problem}
\alias{wbt_travelling_salesman_problem}
\title{Travelling salesman problem}
\usage{
wbt_travelling_salesman_problem(
  input,
  output,
  duration = 60,
  wd = NULL,
  verbose_mode = NULL,
  compress_rasters = NULL,
  command_only = FALSE
)
}
\arguments{
\item{input}{Name of the input points shapefile.}

\item{output}{Name of the output lines shapefile.}

\item{duration}{Maximum duration, in seconds.}

\item{wd}{Changes the working directory. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_wd()} for details.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_verbose()} for details.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_compress_rasters()} for details.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool. Default: \code{FALSE}.}
}
\value{
Returns the tool text outputs.
}
\description{
Finds approximate solutions to travelling salesman problems, the goal of which is to identify the shortest route connecting a set of locations.
}
\keyword{GISAnalysis}
