% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{wbt_turning_bands_simulation}
\alias{wbt_turning_bands_simulation}
\title{Turning bands simulation}
\usage{
wbt_turning_bands_simulation(
  base,
  output,
  range,
  iterations = 1000,
  wd = NULL,
  verbose_mode = NULL,
  compress_rasters = NULL,
  command_only = FALSE
)
}
\arguments{
\item{base}{Input base raster file.}

\item{output}{Output file.}

\item{range}{The field's range, in xy-units, related to the extent of spatial autocorrelation.}

\item{iterations}{The number of iterations.}

\item{wd}{Changes the working directory. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_wd()} for details.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_verbose()} for details.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_compress_rasters()} for details.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool. Default: \code{FALSE}.}
}
\value{
Returns the tool text outputs.
}
\description{
Creates an image containing random values based on a turning-bands simulation.
}
\keyword{MathandStatsTools}
