% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_union}
\alias{wbt_union}
\title{Union}
\usage{
wbt_union(
  input,
  overlay,
  output,
  snap = 0,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{input}{Input vector file.}

\item{overlay}{Input overlay vector file.}

\item{output}{Output vector file.}

\item{snap}{Snap tolerance.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
Splits vector layers at their overlaps, creating a layer containing all the portions from both input and overlay layers.
}
