% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{wbt_unsharp_masking}
\alias{wbt_unsharp_masking}
\title{Unsharp masking}
\usage{
wbt_unsharp_masking(
  input,
  output,
  sigma = 0.75,
  amount = 100,
  threshold = 0,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{sigma}{Standard deviation distance in pixels.}

\item{amount}{A percentage and controls the magnitude of each overshoot.}

\item{threshold}{Controls the minimal brightness change that will be sharpened.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
An image sharpening technique that enhances edges.
}
