% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_update_nodata_cells}
\alias{wbt_update_nodata_cells}
\title{Update nodata cells}
\usage{
wbt_update_nodata_cells(
  input1,
  input2,
  output,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{input1}{Input raster file 1.}

\item{input2}{Input raster file 2; update layer.}

\item{output}{Output raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
Replaces the NoData values in an input raster with the corresponding values contained in a second update layer.
}
