% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{wbt_upslope_depression_storage}
\alias{wbt_upslope_depression_storage}
\title{Upslope depression storage}
\usage{
wbt_upslope_depression_storage(dem, output, verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Estimates the average upslope depression storage depth.
}
