% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{wbt_user_defined_weights_filter}
\alias{wbt_user_defined_weights_filter}
\title{User ined weights filter}
\usage{
wbt_user_defined_weights_filter(
  input,
  weights,
  output,
  center = "center",
  normalize = FALSE,
  verbose_mode = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{weights}{Input weights file.}

\item{output}{Output raster file.}

\item{center}{Kernel center cell; options include 'center', 'upper-left', 'upper-right', 'lower-left', 'lower-right'.}

\item{normalize}{Normalize kernel weights? This can reduce edge effects and lessen the impact of data gaps (nodata) but is not suited when the kernel weights sum to zero.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a user-defined weights filter on an image.
}
