% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{wbt_vector_lines_to_raster}
\alias{wbt_vector_lines_to_raster}
\title{Vector lines to raster}
\usage{
wbt_vector_lines_to_raster(
  input,
  output,
  field = "FID",
  nodata = TRUE,
  cell_size = NULL,
  base = NULL,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{input}{Input vector lines file.}

\item{output}{Output raster file.}

\item{field}{Input field name in attribute table.}

\item{nodata}{Background value to set to NoData. Without this flag, it will be set to 0.0.}

\item{cell_size}{Optionally specified cell size of output raster. Not used when base raster is specified.}

\item{base}{Optionally specified input base raster file. Not used when a cell size is specified.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
Converts a vector containing polylines into a raster.
}
