% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_network_analysis.R
\name{wbt_vector_stream_network_analysis}
\alias{wbt_vector_stream_network_analysis}
\title{Vector stream network analysis}
\usage{
wbt_vector_stream_network_analysis(
  streams,
  dem,
  output,
  cutting_height = 10,
  snap = 0.1,
  wd = NULL,
  verbose_mode = NULL,
  compress_rasters = NULL,
  command_only = FALSE
)
}
\arguments{
\item{streams}{Name of the input streams vector file.}

\item{dem}{Name of the input DEM raster file.}

\item{output}{Name of the output lines shapefile.}

\item{cutting_height}{Maximum ridge-cutting height (z units).}

\item{snap}{Snap distance, in xy units (metres).}

\item{wd}{Changes the working directory. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_wd()} for details.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_verbose()} for details.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_compress_rasters()} for details.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool. Default: \code{FALSE}.}
}
\value{
Returns the tool text outputs.
}
\description{
This tool performs common stream network analysis operations on an input vector stream file.
}
\keyword{StreamNetworkAnalysis}
