% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbt.R
\name{wbt_verbose}
\alias{wbt_verbose}
\title{Check verbose options for WhiteboxTools.}
\usage{
wbt_verbose(verbose = NULL)
}
\arguments{
\item{verbose}{Default: \code{NULL}; if logical, set the package option \code{whitebox.verbose} to specified value}
}
\value{
logical; defaults to result of \code{interactive()}
}
\description{
Return \code{TRUE} if "R_WHITEBOX_VERBOSE" system environment variable or "whitebox.verbose" option is set to \code{TRUE}. Returns \code{TRUE} when R is being used interactively and \code{FALSE} otherwise.
}
\examples{
 wbt_verbose()
 
\dontrun{
 Sys.setenv("R_WHITEBOX_VERBOSE" = FALSE)
 wbt_verbose()
 
 options(whitebox.verbose = FALSE)
 wbt_verbose()
}
}
