% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbt.R
\name{wbt_version}
\alias{wbt_version}
\title{Retrieves the version information for WhiteboxTools.}
\usage{
wbt_version()
}
\value{
Returns the version information for WhiteboxTools as an R character vector.
}
\description{
Retrieves the version information for WhiteboxTools.
}
\examples{
wbt_version()
}
