% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbt.R
\name{wbt_view_code}
\alias{wbt_view_code}
\title{Source code for a specific tool.}
\usage{
wbt_view_code(tool_name, viewer = TRUE)
}
\arguments{
\item{tool_name}{Name of the tool.}

\item{viewer}{Show source code in browser? default: \code{TRUE}}
}
\value{
Returns a GitHub URL to view the source code of the tool.
}
\description{
Opens a web browser to view the source code for a specific tool on the projects source code repository.
}
\examples{
\dontrun{
wbt_view_code("breach_depressions")
}
}
