% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_viewshed}
\alias{wbt_viewshed}
\title{Viewshed}
\usage{
wbt_viewshed(
  dem,
  stations,
  output,
  height = 2,
  wd = NULL,
  verbose_mode = FALSE
)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{stations}{Input viewing station vector file.}

\item{output}{Output raster file.}

\item{height}{Viewing station height, in z units.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Identifies the viewshed for a point or set of points.
}
