% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{wbt_xor}
\alias{wbt_xor}
\title{Xor}
\usage{
wbt_xor(input1, input2, output, verbose_mode = FALSE)
}
\arguments{
\item{input1}{Input raster file.}

\item{input2}{Input raster file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a logical XOR operator on two Boolean raster images.
}
