% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{wbt_zonal_statistics}
\alias{wbt_zonal_statistics}
\title{Zonal statistics
Extracts descriptive statistics for a group of patches in a raster.}
\usage{
wbt_zonal_statistics(
  input,
  features,
  output = NULL,
  stat = "mean",
  out_table = NULL,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input data raster file.}

\item{features}{Input feature definition raster file.}

\item{output}{Output raster file.}

\item{stat}{Statistic to extract, including 'mean', 'median', 'minimum', 'maximum', 'range', 'standard deviation', and 'total'.}

\item{out_table}{Output HTML Table file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Zonal statistics
Extracts descriptive statistics for a group of patches in a raster.
}
\keyword{MathandStatsTools}
