% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{weighted_sum}
\alias{weighted_sum}
\title{Weighted sum}
\usage{
weighted_sum(inputs, weights, output, verbose_mode = TRUE)
}
\arguments{
\item{inputs}{Input raster files.}

\item{weights}{Weight values, contained in quotes and separated by commas or semicolons.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a weighted-sum overlay on multiple input raster images.
}
