% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbt.R
\name{whitebox_license}
\alias{whitebox_license}
\title{Retrieves the license information for WhiteboxTools.}
\usage{
whitebox_license()
}
\value{
Returns the license information for WhiteboxTools as an R character vector.
}
\description{
Retrieves the license information for WhiteboxTools.
}
\examples{
whitebox_license()
}
