% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{z_scores}
\alias{z_scores}
\title{Z scores}
\usage{
z_scores(input, output, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Standardizes the values in an input raster by converting to z-scores.
}
\examples{
dem <- system.file("extdata", "DEM.tif", package="whitebox")
z_scores(input = dem, output = 'output.tif')
}
