% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{breach_depressions}
\alias{breach_depressions}
\title{Breaches all of the depressions in a DEM}
\usage{
breach_depressions(dem, output, max_depth = NULL, max_length = NULL,
  verbose_mode = TRUE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{max_depth}{Optional maximum breach depth (default is Inf).}

\item{max_length}{Optional maximum breach channel length (in grid cells; default is Inf).}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Breaches all of the depressions in a DEM using Lindsay's (2016) algorithm. This should be preferred over depression filling in most cases.
}
\examples{
dem <- system.file("extdata", "DEM.tif", package="whitebox")
output <- file.path(getwd(), "output.tif")
breach_depressions(dem, output, verbose_mode = TRUE)
}
