% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_whitebox_binary}
\alias{check_whitebox_binary}
\title{Check for WhiteboxTools executable path}
\usage{
check_whitebox_binary(silent = TRUE)
}
\arguments{
\item{silent}{logical. Print help on installation/setting path. Default \code{TRUE}.}
}
\value{
logical if WhiteboxTools executable file exists.
}
\description{
Check for WhiteboxTools executable path
}
\seealso{
\code{\link[=wbt_exe_path]{wbt_exe_path()}}
}
