% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbt.R
\name{install_whitebox}
\alias{install_whitebox}
\alias{wbt_install}
\title{Download and Install WhiteboxTools}
\usage{
install_whitebox(pkg_dir = find.package("whitebox"), force = FALSE)
}
\arguments{
\item{pkg_dir}{default install path is to whitebox package "WBT" folder}

\item{force}{logical. Default \code{FALSE}. Force install?}
}
\value{
Prints out the location of the WhiteboxTools binary, if found. \code{NULL} otherwise.
}
\description{
This function downloads the WhiteboxTools binary if needed. Pre-compiled binaries are only available for download for 64-bit Linux (Ubuntu 20.04), Windows and Mac OS (Intel) platforms. If you need WhiteboxTools for another platform follow the instructions here: \url{https://github.com/jblindsay/whitebox-tools}
}
\examples{
\dontrun{
install_whitebox()
}
}
