% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_aggregate_raster}
\alias{wbt_aggregate_raster}
\title{Aggregate raster
Aggregates a raster to a lower resolution.}
\usage{
wbt_aggregate_raster(
  input,
  output,
  agg_factor = 2,
  type = "mean",
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{agg_factor}{Aggregation factor, in pixels.}

\item{type}{Statistic used to fill output pixels.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Aggregate raster
Aggregates a raster to a lower resolution.
}
\keyword{GISAnalysis}
