% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{wbt_anova}
\alias{wbt_anova}
\title{Anova}
\usage{
wbt_anova(input, features, output, wd = NULL, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{features}{Feature definition (or class) raster.}

\item{output}{Output HTML file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs an analysis of variance (ANOVA) test on a raster dataset.
}
