% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{wbt_atan2}
\alias{wbt_atan2}
\title{Atan2}
\usage{
wbt_atan2(input_y, input_x, output, wd = NULL, verbose_mode = FALSE)
}
\arguments{
\item{input_y}{Input y raster file or constant value (rise).}

\item{input_x}{Input x raster file or constant value (run).}

\item{output}{Output raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Returns the 2-argument inverse tangent (atan2).
}
