% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{wbt_attribute_correlation_neighbourhood_analysis}
\alias{wbt_attribute_correlation_neighbourhood_analysis}
\title{Attribute correlation neighbourhood analysis}
\usage{
wbt_attribute_correlation_neighbourhood_analysis(
  input,
  field1,
  field2,
  radius = NULL,
  min_points = NULL,
  stat = "pearson",
  verbose_mode = FALSE
)
}
\arguments{
\item{input}{Input vector file.}

\item{field1}{First input field name (dependent variable) in attribute table.}

\item{field2}{Second input field name (independent variable) in attribute table.}

\item{radius}{Search Radius (in map units).}

\item{min_points}{Minimum number of points.}

\item{stat}{Correlation type; one of 'pearson' (default) and 'spearman'.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a correlation on two input vector attributes within a neighbourhood search windows.
}
