% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{wbt_attribute_scattergram}
\alias{wbt_attribute_scattergram}
\title{Attribute scattergram}
\usage{
wbt_attribute_scattergram(
  input,
  fieldx,
  fieldy,
  output,
  trendline = FALSE,
  wd = NULL,
  verbose_mode = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{fieldx}{Input field name in attribute table for the x-axis.}

\item{fieldy}{Input field name in attribute table for the y-axis.}

\item{output}{Output HTML file (default name will be based on input file if unspecified).}

\item{trendline}{Draw the trendline.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Creates a scattergram for two field values of a vector's attribute table.
}
