% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{wbt_average_upslope_flowpath_length}
\alias{wbt_average_upslope_flowpath_length}
\title{Average upslope flowpath length}
\usage{
wbt_average_upslope_flowpath_length(
  dem,
  output,
  wd = NULL,
  verbose_mode = FALSE
)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Measures the average length of all upslope flowpaths draining each grid cell.
}
