% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{wbt_bilateral_filter}
\alias{wbt_bilateral_filter}
\title{Bilateral filter}
\usage{
wbt_bilateral_filter(
  input,
  output,
  sigma_dist = 0.75,
  sigma_int = 1,
  wd = NULL,
  verbose_mode = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{sigma_dist}{Standard deviation in distance in pixels.}

\item{sigma_int}{Standard deviation in intensity in pixels.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
A bilateral filter is an edge-preserving smoothing filter introduced by Tomasi and Manduchi (1998).
}
