% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{wbt_breach_depressions}
\alias{wbt_breach_depressions}
\title{Breach depressions}
\usage{
wbt_breach_depressions(dem, output, max_depth = NULL,
  max_length = NULL, flat_increment = NULL, fill_pits = FALSE,
  verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{max_depth}{Optional maximum breach depth (default is Inf).}

\item{max_length}{Optional maximum breach channel length (in grid cells; default is Inf).}

\item{flat_increment}{Optional elevation increment applied to flat areas.}

\item{fill_pits}{Optional flag indicating whether to fill single-cell pits.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Breaches all of the depressions in a DEM using Lindsay's (2016) algorithm. This should be preferred over depression filling in most cases.
}
