% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{wbt_breach_single_cell_pits}
\alias{wbt_breach_single_cell_pits}
\title{Breach single cell pits}
\usage{
wbt_breach_single_cell_pits(dem, output, wd = NULL, verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Removes single-cell pits from an input DEM by breaching.
}
