% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_breakline_mapping}
\alias{wbt_breakline_mapping}
\title{Breakline mapping}
\usage{
wbt_breakline_mapping(
  dem,
  output,
  threshold = 2,
  min_length = 3,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{dem}{Name of the input raster image file.}

\item{output}{Name of the output vector lines file.}

\item{threshold}{Threshold value (0 - infinity but typically 1 to 5 works well).}

\item{min_length}{Minimum line length, in grid cells.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
This tool maps breaklines from an input DEM.
}
