% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_buffer_raster}
\alias{wbt_buffer_raster}
\title{Buffer raster}
\usage{
wbt_buffer_raster(input, output, size, gridcells = FALSE,
  verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{size}{Buffer size.}

\item{gridcells}{Optional flag to indicate that the 'size' threshold should be measured in grid cells instead of the default map units.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Maps a distance-based buffer around each non-background (non-zero/non-nodata) grid cell in an input image.
}
