% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{wbt_burn_streams_at_roads}
\alias{wbt_burn_streams_at_roads}
\title{Burn streams at roads}
\usage{
wbt_burn_streams_at_roads(
  dem,
  streams,
  roads,
  output,
  width = NULL,
  wd = NULL,
  verbose_mode = FALSE
)
}
\arguments{
\item{dem}{Input raster digital elevation model (DEM) file.}

\item{streams}{Input vector streams file.}

\item{roads}{Input vector roads file.}

\item{output}{Output raster file.}

\item{width}{Maximum road embankment width, in map units.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Burns-in streams at the sites of road embankments.
}
