% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_circular_variance_of_aspect}
\alias{wbt_circular_variance_of_aspect}
\title{Circular variance of aspect
Calculates the circular variance of aspect at a scale for a DEM.}
\usage{
wbt_circular_variance_of_aspect(
  dem,
  output,
  filter = 11,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{filter}{Size of the filter kernel.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Circular variance of aspect
Calculates the circular variance of aspect at a scale for a DEM.
}
\keyword{GeomorphometricAnalysis}
