% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_classify_buildings_in_lidar}
\alias{wbt_classify_buildings_in_lidar}
\title{Classify buildings in lidar
Reclassifies a LiDAR points that lie within vector building footprints.}
\usage{
wbt_classify_buildings_in_lidar(
  input,
  buildings,
  output,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input LiDAR file.}

\item{buildings}{Input vector polygons file.}

\item{output}{Output LiDAR file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Classify buildings in lidar
Reclassifies a LiDAR points that lie within vector building footprints.
}
\keyword{LiDARTools}
