% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_classify_buildings_in_lidar}
\alias{wbt_classify_buildings_in_lidar}
\title{Classify buildings in lidar}
\usage{
wbt_classify_buildings_in_lidar(input, buildings, output, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input LiDAR file.}

\item{buildings}{Input vector polygons file.}

\item{output}{Output LiDAR file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Reclassifies a LiDAR points that lie within vector building footprints.
}
