% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis_overlay.R
\name{wbt_clip_raster_to_polygon}
\alias{wbt_clip_raster_to_polygon}
\title{Clip raster to polygon}
\usage{
wbt_clip_raster_to_polygon(
  input,
  polygons,
  output,
  maintain_dimensions = TRUE,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{polygons}{Input vector polygons file.}

\item{output}{Output raster file.}

\item{maintain_dimensions}{Maintain input raster dimensions?.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Clips a raster to a vector polygon.
}
