% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_construct_vector_tin}
\alias{wbt_construct_vector_tin}
\title{Construct vector tin}
\usage{
wbt_construct_vector_tin(input, output, field = NULL, use_z = FALSE,
  max_triangle_edge_length = NULL, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input vector points file.}

\item{output}{Output vector polygon file.}

\item{field}{Input field name in attribute table.}

\item{use_z}{Use the 'z' dimension of the Shapefile's geometry instead of an attribute field?.}

\item{max_triangle_edge_length}{Optional maximum triangle edge length; triangles larger than this size will not be gridded.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Creates a vector triangular irregular network (TIN) for a set of vector points.
}
