% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{wbt_correct_vignetting}
\alias{wbt_correct_vignetting}
\title{Correct vignetting}
\usage{
wbt_correct_vignetting(
  input,
  pp,
  output,
  focal_length = 304.8,
  image_width = 228.6,
  n = 4,
  verbose_mode = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{pp}{Input principal point file.}

\item{output}{Output raster file.}

\item{focal_length}{Camera focal length, in millimeters.}

\item{image_width}{Distance between photograph edges, in millimeters.}

\item{n}{The 'n' parameter.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Corrects the darkening of images towards corners.
}
