% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis_enhancement.R
\name{wbt_correct_vignetting}
\alias{wbt_correct_vignetting}
\title{Correct vignetting}
\usage{
wbt_correct_vignetting(
  input,
  pp,
  output,
  focal_length = 304.8,
  image_width = 228.6,
  n = 4,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{pp}{Input principal point file.}

\item{output}{Output raster file.}

\item{focal_length}{Camera focal length, in millimeters.}

\item{image_width}{Distance between photograph edges, in millimeters.}

\item{n}{The 'n' parameter.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Corrects the darkening of images towards corners.
}
