% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_cost_allocation}
\alias{wbt_cost_allocation}
\title{Cost allocation}
\usage{
wbt_cost_allocation(
  source,
  backlink,
  output,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{source}{Input source raster file.}

\item{backlink}{Input backlink raster file generated by the cost-distance tool.}

\item{output}{Output raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
Identifies the source cell to which each grid cell is connected by a least-cost pathway in a cost-distance analysis.
}
