% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_cost_distance}
\alias{wbt_cost_distance}
\title{Cost distance}
\usage{
wbt_cost_distance(
  source,
  cost,
  out_accum,
  out_backlink,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{source}{Input source raster file.}

\item{cost}{Input cost (friction) raster file.}

\item{out_accum}{Output cost accumulation raster file.}

\item{out_backlink}{Output backlink raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs cost-distance accumulation on a cost surface and a group of source cells.
}
