% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{wbt_create_colour_composite}
\alias{wbt_create_colour_composite}
\title{Create colour composite
Creates a colour-composite image from three bands of multispectral imagery.}
\usage{
wbt_create_colour_composite(
  red,
  green,
  blue,
  output,
  opacity = NULL,
  enhance = TRUE,
  zeros = FALSE,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{red}{Input red band image file.}

\item{green}{Input green band image file.}

\item{blue}{Input blue band image file.}

\item{output}{Output colour composite file.}

\item{opacity}{Input opacity band image file (optional).}

\item{enhance}{Optional flag indicating whether a balance contrast enhancement is performed.}

\item{zeros}{Optional flag to indicate if zeros are nodata values.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Create colour composite
Creates a colour-composite image from three bands of multispectral imagery.
}
\keyword{ImageProcessingTools}
