% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_create_plane}
\alias{wbt_create_plane}
\title{Create plane}
\usage{
wbt_create_plane(
  base,
  output,
  gradient = 15,
  aspect = 90,
  constant = 0,
  wd = NULL,
  verbose_mode = NULL,
  compress_rasters = NULL,
  command_only = FALSE
)
}
\arguments{
\item{base}{Input base raster file.}

\item{output}{Output raster file.}

\item{gradient}{Slope gradient in degrees (-85.0 to 85.0).}

\item{aspect}{Aspect (direction) in degrees clockwise from north (0.0-360.0).}

\item{constant}{Constant value.}

\item{wd}{Changes the working directory. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_wd()} for details.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_verbose()} for details.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_compress_rasters()} for details.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool. Default: \code{FALSE}.}
}
\value{
Returns the tool text outputs.
}
\description{
Creates a raster image based on the equation for a simple plane.
}
\keyword{GISAnalysis}
