% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_create_rectangular_vector_grid}
\alias{wbt_create_rectangular_vector_grid}
\title{Create rectangular vector grid}
\usage{
wbt_create_rectangular_vector_grid(input, output, width, height,
  xorig = 0, yorig = 0, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input base file.}

\item{output}{Output vector polygon file.}

\item{width}{The grid cell width.}

\item{height}{The grid cell height.}

\item{xorig}{The grid origin x-coordinate.}

\item{yorig}{The grid origin y-coordinate.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Creates a rectangular vector grid.
}
