% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{wbt_depth_in_sink}
\alias{wbt_depth_in_sink}
\title{Depth in sink}
\usage{
wbt_depth_in_sink(dem, output, zero_background = FALSE, verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{zero_background}{Flag indicating whether the background value of zero should be used.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Measures the depth of sinks (depressions) in a DEM.
}
