% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_directional_relief}
\alias{wbt_directional_relief}
\title{Directional relief}
\usage{
wbt_directional_relief(dem, output, azimuth = 0, max_dist = NULL,
  verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{azimuth}{Wind azimuth in degrees.}

\item{max_dist}{Optional maximum search distance (unspecified if none; in xy units).}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates relief for cells in an input DEM for a specified direction.
}
