% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_dissolve}
\alias{wbt_dissolve}
\title{Dissolve}
\usage{
wbt_dissolve(
  input,
  output,
  field = NULL,
  snap = 0,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{input}{Input vector file.}

\item{output}{Output vector file.}

\item{field}{Dissolve field attribute (optional).}

\item{snap}{Snap tolerance.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
Removes the interior, or shared, boundaries within a vector polygon coverage.
}
