% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{wbt_downslope_flowpath_length}
\alias{wbt_downslope_flowpath_length}
\title{Downslope flowpath length}
\usage{
wbt_downslope_flowpath_length(
  d8_pntr,
  output,
  watersheds = NULL,
  weights = NULL,
  esri_pntr = FALSE,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{d8_pntr}{Input D8 pointer raster file.}

\item{output}{Output raster file.}

\item{watersheds}{Optional input watershed raster file.}

\item{weights}{Optional input weights raster file.}

\item{esri_pntr}{D8 pointer uses the ESRI style scheme.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the downslope flowpath length from each cell to basin outlet.
}
\keyword{HydrologicalAnalysis}
