% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis_patch_shape.R
\name{wbt_edge_proportion}
\alias{wbt_edge_proportion}
\title{Edge proportion
Calculate the proportion of cells in a raster polygon that are edge cells.}
\usage{
wbt_edge_proportion(
  input,
  output,
  output_text = FALSE,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{output_text}{flag indicating whether a text report should also be output.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Edge proportion
Calculate the proportion of cells in a raster polygon that are edge cells.
}
\keyword{GISAnalysisPatchShapeTools}
