% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_elev_percentile}
\alias{wbt_elev_percentile}
\title{Elev percentile}
\usage{
wbt_elev_percentile(dem, output, filterx = 11, filtery = 11,
  sig_digits = 2, verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{filterx}{Size of the filter kernel in the x-direction.}

\item{filtery}{Size of the filter kernel in the y-direction.}

\item{sig_digits}{Number of significant digits.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the elevation percentile raster from a DEM.
}
