% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_elev_relative_to_watershed_min_max}
\alias{wbt_elev_relative_to_watershed_min_max}
\title{Elev relative to watershed min max}
\usage{
wbt_elev_relative_to_watershed_min_max(
  dem,
  watersheds,
  output,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{watersheds}{Input raster watersheds file.}

\item{output}{Output raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the elevation of a location relative to the minimum and maximum elevations in a watershed.
}
\keyword{GeomorphometricAnalysis}
