% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_eliminate_coincident_points}
\alias{wbt_eliminate_coincident_points}
\title{Eliminate coincident points}
\usage{
wbt_eliminate_coincident_points(input, output, tolerance,
  verbose_mode = FALSE)
}
\arguments{
\item{input}{Input vector file.}

\item{output}{Output vector polygon file.}

\item{tolerance}{The distance tolerance for points.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Removes any coincident, or nearly coincident, points from a vector points file.
}
