% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_eliminate_coincident_points}
\alias{wbt_eliminate_coincident_points}
\title{Eliminate coincident points}
\usage{
wbt_eliminate_coincident_points(
  input,
  output,
  tolerance,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input vector file.}

\item{output}{Output vector points file.}

\item{tolerance}{The distance tolerance for points.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Removes any coincident, or nearly coincident, points from a vector points file.
}
