% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{wbt_emboss_filter}
\alias{wbt_emboss_filter}
\title{Emboss filter}
\usage{
wbt_emboss_filter(
  input,
  output,
  direction = "n",
  clip = 0,
  verbose_mode = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{direction}{Direction of reflection; options include 'n', 's', 'e', 'w', 'ne', 'se', 'nw', 'sw'.}

\item{clip}{Optional amount to clip the distribution tails by, in percent.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs an emboss filter on an image, similar to a hillshade operation.
}
