% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_erase_polygon_from_lidar}
\alias{wbt_erase_polygon_from_lidar}
\title{Erase polygon from lidar
Erases (cuts out) a vector polygon or polygons from a LiDAR point cloud.}
\usage{
wbt_erase_polygon_from_lidar(
  input,
  polygons,
  output,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input LiDAR file.}

\item{polygons}{Input vector polygons file.}

\item{output}{Output LiDAR file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Erase polygon from lidar
Erases (cuts out) a vector polygon or polygons from a LiDAR point cloud.
}
\keyword{LiDARTools}
