% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis_classification.R
\name{wbt_evaluate_training_sites}
\alias{wbt_evaluate_training_sites}
\title{Evaluate training sites
This tool can be used to inspect the overlap in spectral signatures of training sites for various classes.}
\usage{
wbt_evaluate_training_sites(
  inputs,
  polys,
  field,
  output,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{inputs}{Name of the input band images.}

\item{polys}{Name of the input training site polygons shapefile.}

\item{field}{Name of the attribute containing class name data.}

\item{output}{Name of the output report file (*.html).}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Evaluate training sites
This tool can be used to inspect the overlap in spectral signatures of training sites for various classes.
}
\keyword{ImageProcessingToolsClassification}
