% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_extend_vector_lines}
\alias{wbt_extend_vector_lines}
\title{Extend vector lines}
\usage{
wbt_extend_vector_lines(
  input,
  output,
  dist,
  extend = "both ends",
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input vector polyline file.}

\item{output}{Output vector polyline file.}

\item{dist}{The distance to extend.}

\item{extend}{Extend direction, 'both ends' (default), 'line start', 'line end'.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Extends vector lines by a specified distance.
}
