% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_fill_missing_data}
\alias{wbt_fill_missing_data}
\title{Fill missing data}
\usage{
wbt_fill_missing_data(
  input,
  output,
  filter = 11,
  weight = 2,
  no_edges = TRUE,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{filter}{Filter size (cells).}

\item{weight}{IDW weight value.}

\item{no_edges}{Optional flag indicating whether to exclude NoData cells in edge regions.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Fills NoData holes in a DEM.
}
