% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_fill_missing_data}
\alias{wbt_fill_missing_data}
\title{Fill missing data}
\usage{
wbt_fill_missing_data(
  input,
  output,
  filter = 11,
  weight = 2,
  wd = NULL,
  verbose_mode = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{filter}{Filter size (cells).}

\item{weight}{IDW weight value.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Fills NoData holes in a DEM.
}
