% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_filter_raster_features_by_area}
\alias{wbt_filter_raster_features_by_area}
\title{Filter raster features by area}
\usage{
wbt_filter_raster_features_by_area(
  input,
  output,
  threshold,
  background = "zero",
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{threshold}{Remove features with fewer grid cells than this threshold value.}

\item{background}{Background value.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Removes small-area features from a raster.
}
\keyword{GISAnalysis}
