% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_find_ridges}
\alias{wbt_find_ridges}
\title{Find ridges}
\usage{
wbt_find_ridges(
  dem,
  output,
  line_thin = TRUE,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{line_thin}{Optional flag indicating whether post-processing line-thinning should be performed.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
Identifies potential ridge and peak grid cells.
}
