% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{wbt_fix_dangling_arcs}
\alias{wbt_fix_dangling_arcs}
\title{Fix dangling arcs
This tool fixes undershot and overshot arcs, two common topological errors, in an input vector lines file.}
\usage{
wbt_fix_dangling_arcs(
  input,
  output,
  dist = "",
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Name of the input lines vector file.}

\item{output}{Name of the output lines vector file.}

\item{dist}{Snap distance, in xy units (metres).}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Fix dangling arcs
This tool fixes undershot and overshot arcs, two common topological errors, in an input vector lines file.
}
\keyword{DataTools}
