% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{wbt_flip_image}
\alias{wbt_flip_image}
\title{Flip image}
\usage{
wbt_flip_image(input, output, direction = "vertical",
  verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{direction}{Direction of reflection; options include 'v' (vertical), 'h' (horizontal), and 'b' (both).}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Reflects an image in the vertical or horizontal axis.
}
