% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{wbt_flow_length_diff}
\alias{wbt_flow_length_diff}
\title{Flow length diff}
\usage{
wbt_flow_length_diff(d8_pntr, output, esri_pntr = FALSE, verbose_mode = FALSE)
}
\arguments{
\item{d8_pntr}{Input D8 pointer raster file.}

\item{output}{Output raster file.}

\item{esri_pntr}{D8 pointer uses the ESRI style scheme.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the local maximum absolute difference in downslope flowpath length, useful in mapping drainage divides and ridges.
}
