% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbt.R
\name{wbt_help}
\alias{wbt_help}
\title{Help description for WhiteboxTools.}
\usage{
wbt_help()
}
\value{
Returns the help description for WhiteboxTools as an R character vector.
}
\description{
Help description for WhiteboxTools.
}
\examples{
\dontrun{
wbt_help()
}
}
