% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_highest_position}
\alias{wbt_highest_position}
\title{Highest position}
\usage{
wbt_highest_position(inputs, output, wd = NULL, verbose_mode = FALSE)
}
\arguments{
\item{inputs}{Input raster files.}

\item{output}{Output raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Identifies the stack position of the maximum value within a raster stack on a cell-by-cell basis.
}
