% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_hillshade}
\alias{wbt_hillshade}
\title{Hillshade}
\usage{
wbt_hillshade(
  dem,
  output,
  azimuth = 315,
  altitude = 30,
  zfactor = 1,
  verbose_mode = FALSE
)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{azimuth}{Illumination source azimuth in degrees.}

\item{altitude}{Illumination source altitude in degrees.}

\item{zfactor}{Optional multiplier for when the vertical and horizontal units are not the same.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates a hillshade raster from an input DEM.
}
