% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{wbt_histogram_matching}
\alias{wbt_histogram_matching}
\title{Histogram matching}
\usage{
wbt_histogram_matching(
  input,
  histo_file,
  output,
  wd = NULL,
  verbose_mode = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{histo_file}{Input reference probability distribution function (pdf) text file.}

\item{output}{Output raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Alters the statistical distribution of a raster image matching it to a specified PDF.
}
