% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis_enhancement.R
\name{wbt_histogram_matching}
\alias{wbt_histogram_matching}
\title{Histogram matching
Alters the statistical distribution of a raster image matching it to a specified PDF.}
\usage{
wbt_histogram_matching(
  input,
  histo_file,
  output,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{histo_file}{Input reference probability distribution function (pdf) text file.}

\item{output}{Output raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Histogram matching
Alters the statistical distribution of a raster image matching it to a specified PDF.
}
\keyword{ImageProcessingToolsImageEnhancement}
