% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis_enhancement.R
\name{wbt_histogram_matching_two_images}
\alias{wbt_histogram_matching_two_images}
\title{Histogram matching two images}
\usage{
wbt_histogram_matching_two_images(
  input1,
  input2,
  output,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input1}{Input raster file to modify.}

\item{input2}{Input reference raster file.}

\item{output}{Output raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Alters the cumulative distribution function of a raster image to that of another image.
}
\keyword{ImageProcessingToolsImageEnhancement}
