% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_hole_proportion}
\alias{wbt_hole_proportion}
\title{Hole proportion}
\usage{
wbt_hole_proportion(input, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input vector polygon file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the proportion of the total area of a polygon's holes relative to the area of the polygon's hull.
}
