% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_hypsometric_analysis}
\alias{wbt_hypsometric_analysis}
\title{Hypsometric analysis}
\usage{
wbt_hypsometric_analysis(
  inputs,
  output,
  watershed = NULL,
  wd = NULL,
  verbose_mode = FALSE
)
}
\arguments{
\item{inputs}{Input DEM files.}

\item{output}{Output HTML file (default name will be based on input file if unspecified).}

\item{watershed}{Input watershed files (optional).}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates a hypsometric curve for one or more DEMs.
}
