% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_idw_interpolation}
\alias{wbt_idw_interpolation}
\title{Idw interpolation}
\usage{
wbt_idw_interpolation(
  input,
  field,
  output,
  use_z = FALSE,
  weight = 2,
  radius = NULL,
  min_points = NULL,
  cell_size = NULL,
  base = NULL,
  wd = NULL,
  verbose_mode = FALSE
)
}
\arguments{
\item{input}{Input vector Points file.}

\item{field}{Input field name in attribute table.}

\item{output}{Output raster file.}

\item{use_z}{Use z-coordinate instead of field?.}

\item{weight}{IDW weight value.}

\item{radius}{Search Radius in map units.}

\item{min_points}{Minimum number of points.}

\item{cell_size}{Optionally specified cell size of output raster. Not used when base raster is specified.}

\item{base}{Optionally specified input base raster file. Not used when a cell size is specified.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Interpolates vector points into a raster surface using an inverse-distance weighted scheme.
}
