% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{wbt_image_autocorrelation}
\alias{wbt_image_autocorrelation}
\title{Image autocorrelation}
\usage{
wbt_image_autocorrelation(
  inputs,
  output,
  contiguity = "Rook",
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{inputs}{Input raster files.}

\item{output}{Output HTML file (default name will be based on input file if unspecified).}

\item{contiguity}{Contiguity type.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs Moran's I analysis on two or more input images.
}
\keyword{MathandStatsTools}
