% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{wbt_image_stack_profile}
\alias{wbt_image_stack_profile}
\title{Image stack profile}
\usage{
wbt_image_stack_profile(inputs, points, output, verbose_mode = FALSE)
}
\arguments{
\item{inputs}{Input multispectral image files.}

\item{points}{Input vector points file.}

\item{output}{Output HTML file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Plots an image stack profile (i.e. signature) for a set of points and multispectral images.
}
