% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{wbt_impoundment_size_index}
\alias{wbt_impoundment_size_index}
\title{Impoundment size index}
\usage{
wbt_impoundment_size_index(dem, output, damlength, out_type = "depth",
  verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output file.}

\item{damlength}{Maximum length of thr dam.}

\item{out_type}{Output type; one of 'depth' (default), 'volume', and 'area'.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the impoundment size resulting from damming a DEM.
}
