% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{wbt_impoundment_size_index}
\alias{wbt_impoundment_size_index}
\title{Impoundment size index}
\usage{
wbt_impoundment_size_index(
  dem,
  output,
  damlength,
  out_type = "mean depth",
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output file.}

\item{damlength}{Maximum length of the dam.}

\item{out_type}{Output type; one of 'mean depth' (default), 'volume', 'area', 'max depth'.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the impoundment size resulting from damming a DEM.
}
