% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbt.R
\name{wbt_init}
\alias{wbt_init}
\title{Initialize whitebox}
\usage{
wbt_init()
}
\value{
Prints out the location of the WhiteboxTools binary
}
\description{
This function downloads the WhiteboxTools binary if needed.
}
\examples{
\dontrun{
wbt_init()
}
}
