% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbt.R
\name{wbt_init}
\alias{wbt_init}
\title{Initialize WhiteboxTools}
\usage{
wbt_init(exe_path = wbt_exe_path(shell_quote = FALSE))
}
\arguments{
\item{exe_path}{Default \code{exe_path} is result of \code{wbt_exe_path()} which checks a few user-settable options before defaulting to the package installation directory sub-directory "WBT". May be overridden if a custom path is needed.}
}
\value{
logical; \code{TRUE} if binary file is found at \code{exe_path}
}
\description{
This function is called to check if a suitable whitebox installation is present. If the file exists the package option \code{whitebox.exe_path} is set.
}
\examples{
\dontrun{

# set path to binary as R global option
options(whitebox.exe_path = "C:\\\\path\\\\to\\\\whitebox_tools.exe")

# set path to binary as an environment variable
Sys.setenv(R_WHITEBOX_EXE_PATH = "/path/to/whitebox_tools")

# or set path to binary as an argument
wbt_init(exe_path = "not/a/valid/path/whitebox_tools.exe")

}
}
