% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{wbt_insert_dams}
\alias{wbt_insert_dams}
\title{Insert dams}
\usage{
wbt_insert_dams(
  dem,
  dam_pts,
  output,
  damlength,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{dam_pts}{Input vector dam points file.}

\item{output}{Output file.}

\item{damlength}{Maximum length of the dam.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the impoundment size resulting from damming a DEM.
}
