% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis_overlay.R
\name{wbt_intersect}
\alias{wbt_intersect}
\title{Intersect}
\usage{
wbt_intersect(
  input,
  overlay,
  output,
  snap = 0,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input vector file.}

\item{overlay}{Input overlay vector file.}

\item{output}{Output vector file.}

\item{snap}{Snap tolerance.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Identifies the parts of features in common between two input vector layers.
}
