% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{wbt_ks_test_for_normality}
\alias{wbt_ks_test_for_normality}
\title{Ks test for normality}
\usage{
wbt_ks_test_for_normality(
  input,
  output,
  num_samples = NULL,
  verbose_mode = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output HTML file.}

\item{num_samples}{Number of samples. Leave blank to use whole image.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Evaluates whether the values in a raster are normally distributed.
}
