% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{wbt_laplacian_filter}
\alias{wbt_laplacian_filter}
\title{Laplacian filter}
\usage{
wbt_laplacian_filter(
  input,
  output,
  variant = "3x3(1)",
  clip = 0,
  wd = NULL,
  verbose_mode = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{variant}{Optional variant value. Options include 3x3(1), 3x3(2), 3x3(3), 3x3(4), 5x5(1), and 5x5(2) (default is 3x3(1)).}

\item{clip}{Optional amount to clip the distribution tails by, in percent.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a Laplacian filter on an image.
}
