% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_las_to_zlidar}
\alias{wbt_las_to_zlidar}
\title{Las to zlidar}
\usage{
wbt_las_to_zlidar(
  inputs = NULL,
  outdir = NULL,
  compress = "brotli",
  level = 5,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{inputs}{Input LAS files.}

\item{outdir}{Output directory into which zlidar files are created. If unspecified, it is assumed to be the same as the inputs.}

\item{compress}{Compression method, including 'brotli' and 'deflate'.}

\item{level}{Compression level (1-9).}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
Converts one or more LAS files into the zlidar compressed LiDAR data format.
}
