% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_laz_to_las}
\alias{wbt_laz_to_las}
\title{Laz to las}
\usage{
wbt_laz_to_las(
  input,
  output = NULL,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{input}{Name of the input LAZ files (leave blank to use all LAZ files in WorkingDirectory.}

\item{output}{Output LAS file (including extension).}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
This tool converts one or more LAZ files into the LAS format.
}
