% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_laz_to_las}
\alias{wbt_laz_to_las}
\title{Laz to las
This tool converts one or more LAZ files into the LAS format.}
\usage{
wbt_laz_to_las(
  input,
  output = NULL,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Name of the input LAZ files (leave blank to use all LAZ files in WorkingDirectory.}

\item{output}{Output LAS file (including extension).}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Laz to las
This tool converts one or more LAZ files into the LAS format.
}
\keyword{LiDARTools}
