% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbt.R
\name{wbt_license}
\alias{wbt_license}
\title{License information for 'WhiteboxTools'}
\usage{
wbt_license()
}
\value{
Returns the license information for WhiteboxTools as an R character vector.
}
\description{
License information for 'WhiteboxTools'
}
\examples{
\dontrun{
wbt_license()
}
}
\keyword{General}
