% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_lidar_classify_subset}
\alias{wbt_lidar_classify_subset}
\title{Lidar classify subset}
\usage{
wbt_lidar_classify_subset(
  base,
  subset,
  output,
  subset_class,
  nonsubset_class = NULL,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{base}{Input base LiDAR file.}

\item{subset}{Input subset LiDAR file.}

\item{output}{Output LiDAR file.}

\item{subset_class}{Subset point class value (must be 0-18; see LAS specifications).}

\item{nonsubset_class}{Non-subset point class value (must be 0-18; see LAS specifications).}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Classifies the values in one LiDAR point cloud that correspond with points in a subset cloud.
}
\keyword{LiDARTools}
