% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_lidar_colourize}
\alias{wbt_lidar_colourize}
\title{Lidar colourize}
\usage{
wbt_lidar_colourize(
  in_lidar,
  in_image,
  output,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{in_lidar}{Input LiDAR file.}

\item{in_image}{Input colour image file.}

\item{output}{Output LiDAR file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Adds the red-green-blue colour fields of a LiDAR (LAS) file based on an input image.
}
