% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_lidar_eigenvalue_features}
\alias{wbt_lidar_eigenvalue_features}
\title{Lidar eigenvalue features}
\usage{
wbt_lidar_eigenvalue_features(
  input,
  num_neighbours = NULL,
  radius = NULL,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Name of the input LiDAR points.}

\item{num_neighbours}{Number of neighbours used in search.}

\item{radius}{Search distance used in neighbourhood search.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculate eigenvalue-based metrics from a LiDAR point cloud.
}
\keyword{LiDARTools}
