% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_lidar_hex_binning}
\alias{wbt_lidar_hex_binning}
\title{Lidar hex binning}
\usage{
wbt_lidar_hex_binning(
  input,
  output,
  width,
  orientation = "horizontal",
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input base file.}

\item{output}{Output vector polygon file.}

\item{width}{The grid cell width.}

\item{orientation}{Grid Orientation, 'horizontal' or 'vertical'.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Hex-bins a set of LiDAR points.
}
