% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_lidar_histogram}
\alias{wbt_lidar_histogram}
\title{Lidar histogram}
\usage{
wbt_lidar_histogram(
  input,
  output,
  parameter = "elevation",
  clip = 1,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{input}{Input LiDAR file.}

\item{output}{Output HTML file (default name will be based on input file if unspecified).}

\item{parameter}{Parameter; options are 'elevation' (default), 'intensity', 'scan angle', 'class', 'time'.}

\item{clip}{Amount to clip distribution tails (in percent).}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
Creates a histogram of LiDAR data.
}
