% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_lidar_histogram}
\alias{wbt_lidar_histogram}
\title{Lidar histogram}
\usage{
wbt_lidar_histogram(
  input,
  output,
  parameter = "elevation",
  clip = 1,
  verbose_mode = FALSE
)
}
\arguments{
\item{input}{Input LiDAR file.}

\item{output}{Output HTML file (default name will be based on input file if unspecified).}

\item{parameter}{Parameter; options are 'elevation' (default), 'intensity', 'scan angle', 'class'.}

\item{clip}{Amount to clip distribution tails (in percent).}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Creates a histogram of LiDAR data.
}
