% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_lidar_info}
\alias{wbt_lidar_info}
\title{Lidar info}
\usage{
wbt_lidar_info(input, output = NULL, vlr = TRUE, geokeys = TRUE,
  verbose_mode = FALSE)
}
\arguments{
\item{input}{Input LiDAR file.}

\item{output}{Output HTML file for summary report.}

\item{vlr}{Flag indicating whether or not to print the variable length records (VLRs).}

\item{geokeys}{Flag indicating whether or not to print the geokeys.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Prints information about a LiDAR (LAS) dataset, including header, point return frequency, and classification data and information about the variable length records (VLRs) and geokeys.
}
