% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_lidar_thin}
\alias{wbt_lidar_thin}
\title{Lidar thin}
\usage{
wbt_lidar_thin(input, output, resolution = 2, method = "lowest",
  save_filtered = FALSE, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input LiDAR file.}

\item{output}{Output LiDAR file.}

\item{resolution}{The size of the square area used to evaluate nearby points in the LiDAR data.}

\item{method}{Point selection method; options are 'first', 'last', 'lowest' (default), 'highest', 'nearest'.}

\item{save_filtered}{Save filtered points to seperate file?.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Thins a LiDAR point cloud, reducing point density.
}
