% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_lidar_thin_high_density}
\alias{wbt_lidar_thin_high_density}
\title{Lidar thin high density}
\usage{
wbt_lidar_thin_high_density(input, output, density, resolution = 1,
  save_filtered = FALSE, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input LiDAR file.}

\item{output}{Output LiDAR file.}

\item{density}{Max. point density (points / m^3).}

\item{resolution}{Output raster's grid resolution.}

\item{save_filtered}{Save filtered points to seperate file?.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Thins points from high density areas within a LiDAR point cloud.
}
