% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{wbt_line_detection_filter}
\alias{wbt_line_detection_filter}
\title{Line detection filter}
\usage{
wbt_line_detection_filter(
  input,
  output,
  variant = "vertical",
  absvals = FALSE,
  clip = 0,
  wd = NULL,
  verbose_mode = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{variant}{Optional variant value. Options include 'v' (vertical), 'h' (horizontal), '45', and '135' (default is 'v').}

\item{absvals}{Optional flag indicating whether outputs should be absolute values.}

\item{clip}{Optional amount to clip the distribution tails by, in percent.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a line-detection filter on an image.
}
