% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{wbt_list_unique_values}
\alias{wbt_list_unique_values}
\title{List unique values}
\usage{
wbt_list_unique_values(
  input,
  field,
  output,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{field}{Input field name in attribute table.}

\item{output}{Output HTML file (default name will be based on input file if unspecified).}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Lists the unique values contained in a field within a vector's attribute table.
}
\keyword{MathandStatsTools}
