% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_local_quadratic_regression}
\alias{wbt_local_quadratic_regression}
\title{Local quadratic regression
This tool is an implementation of the constrained quadratic regression algorithm using a flexible window size described in Wood (1996).}
\usage{
wbt_local_quadratic_regression(
  dem,
  output,
  filter = 3,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{dem}{Name of the input DEM raster file.}

\item{output}{Name of the output raster file.}

\item{filter}{Edge length of the filter kernel.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Local quadratic regression
This tool is an implementation of the constrained quadratic regression algorithm using a flexible window size described in Wood (1996).
}
\keyword{GeomorphometricAnalysis}
