% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_map_off_terrain_objects}
\alias{wbt_map_off_terrain_objects}
\title{Map off terrain objects}
\usage{
wbt_map_off_terrain_objects(
  dem,
  output,
  max_slope = 40,
  min_size = 1,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{max_slope}{Maximum inter-cell absolute slope.}

\item{min_size}{Minimum feature size, in grid cells.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Maps off-terrain objects in a digital elevation model (DEM).
}
