% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis_overlay.R
\name{wbt_max_absolute_overlay}
\alias{wbt_max_absolute_overlay}
\title{Max absolute overlay}
\usage{
wbt_max_absolute_overlay(
  inputs,
  output,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{inputs}{Input raster files.}

\item{output}{Output raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Evaluates the maximum absolute value for each grid cell from a stack of input rasters.
}
