% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{wbt_merge_table_with_csv}
\alias{wbt_merge_table_with_csv}
\title{Merge table with csv}
\usage{
wbt_merge_table_with_csv(
  input,
  pkey,
  csv,
  fkey,
  import_field = NULL,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input primary vector file (i.e. the table to be modified).}

\item{pkey}{Primary key field.}

\item{csv}{Input CSV file (i.e. source of data to be imported).}

\item{fkey}{Foreign key field.}

\item{import_field}{Imported field (all fields will be imported if not specified).}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Merge a vector's attribute table with a table contained within a CSV text file.
}
