% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_min_downslope_elev_change}
\alias{wbt_min_downslope_elev_change}
\title{Min downslope elev change}
\usage{
wbt_min_downslope_elev_change(dem, output, wd = NULL, verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the minimum downslope change in elevation between a grid cell and its eight downslope neighbors.
}
