% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_minimum_bounding_box}
\alias{wbt_minimum_bounding_box}
\title{Minimum bounding box}
\usage{
wbt_minimum_bounding_box(input, output, criterion = "area",
  features = TRUE, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input vector file.}

\item{output}{Output vector polygon file.}

\item{criterion}{Minimization criterion; options include 'area' (default), 'length', 'width', and 'perimeter'.}

\item{features}{Find the minimum bounding rectangles around each individual vector feature.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Creates a vector minimum bounding rectangle around vector features.
}
