% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_minimum_bounding_circle}
\alias{wbt_minimum_bounding_circle}
\title{Minimum bounding circle}
\usage{
wbt_minimum_bounding_circle(input, output, features = TRUE,
  verbose_mode = FALSE)
}
\arguments{
\item{input}{Input vector file.}

\item{output}{Output vector polygon file.}

\item{features}{Find the minimum bounding circle around each individual vector feature.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Delineates the minimum bounding circle (i.e. smallest enclosing circle) for a group of vectors.
}
