% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_modified_k_means_clustering}
\alias{wbt_modified_k_means_clustering}
\title{Modified k means clustering}
\usage{
wbt_modified_k_means_clustering(
  inputs,
  output,
  out_html = NULL,
  start_clusters = 1000,
  merge_dist = NULL,
  max_iterations = 10,
  class_change = 2,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{inputs}{Input raster files.}

\item{output}{Output raster file.}

\item{out_html}{Output HTML report file.}

\item{start_clusters}{Initial number of clusters.}

\item{merge_dist}{Cluster merger distance.}

\item{max_iterations}{Maximum number of iterations.}

\item{class_change}{Minimum percent of cells changed between iterations before completion.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a modified k-means clustering operation on a multi-spectral dataset.
}
