% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_multiscale_topographic_position_image}
\alias{wbt_multiscale_topographic_position_image}
\title{Multiscale topographic position image}
\usage{
wbt_multiscale_topographic_position_image(
  local,
  meso,
  broad,
  output,
  hillshade = NULL,
  lightness = 1.2,
  wd = NULL,
  verbose_mode = NULL,
  compress_rasters = NULL,
  command_only = FALSE
)
}
\arguments{
\item{local}{Input local-scale topographic position (DEVmax) raster file.}

\item{meso}{Input meso-scale topographic position (DEVmax) raster file.}

\item{broad}{Input broad-scale topographic position (DEVmax) raster file.}

\item{output}{Output raster file.}

\item{hillshade}{Input optional hillshade raster file. Note: a multi-directional (360-degree option) hillshade tends to work best in this application.}

\item{lightness}{Image lightness value (default is 1.2).}

\item{wd}{Changes the working directory. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_wd()} for details.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_verbose()} for details.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_compress_rasters()} for details.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool. Default: \code{FALSE}.}
}
\value{
Returns the tool text outputs.
}
\description{
Creates a multiscale topographic position image from three DEVmax rasters of differing spatial scale ranges.
}
\keyword{GeomorphometricAnalysis}
