% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis_patch_shape.R
\name{wbt_narrowness_index}
\alias{wbt_narrowness_index}
\title{Narrowness index}
\usage{
wbt_narrowness_index(
  input,
  output,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the narrowness of raster polygons.
}
