% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{wbt_new_raster_from_base}
\alias{wbt_new_raster_from_base}
\title{New raster from base}
\usage{
wbt_new_raster_from_base(
  base,
  output,
  value = "nodata",
  data_type = "float",
  verbose_mode = FALSE
)
}
\arguments{
\item{base}{Input base raster file.}

\item{output}{Output raster file.}

\item{value}{Constant value to fill raster with; either 'nodata' or numeric value.}

\item{data_type}{Output raster data type; options include 'double' (64-bit), 'float' (32-bit), and 'integer' (signed 16-bit) (default is 'float').}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Creates a new raster using a base image.
}
