% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_normal_vectors}
\alias{wbt_normal_vectors}
\title{Normal vectors
Calculates normal vectors for points within a LAS file and stores these data (XYZ vector components) in the RGB field.}
\usage{
wbt_normal_vectors(
  input,
  output,
  radius = 1,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input LiDAR file.}

\item{output}{Output LiDAR file.}

\item{radius}{Search Radius.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Normal vectors
Calculates normal vectors for points within a LAS file and stores these data (XYZ vector components) in the RGB field.
}
\keyword{LiDARTools}
