% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_num_upslope_neighbours}
\alias{wbt_num_upslope_neighbours}
\title{Num upslope neighbours}
\usage{
wbt_num_upslope_neighbours(dem, output, verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the number of upslope neighbours to each grid cell in a DEM.
}
