% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis_classification.R
\name{wbt_parallelepiped_classification}
\alias{wbt_parallelepiped_classification}
\title{Parallelepiped classification}
\usage{
wbt_parallelepiped_classification(
  inputs,
  polys,
  field,
  output,
  wd = NULL,
  verbose_mode = NULL,
  compress_rasters = NULL,
  command_only = FALSE
)
}
\arguments{
\item{inputs}{Name of the input band images.}

\item{polys}{Name of the input training site polygons shapefile.}

\item{field}{Name of the attribute containing class name data.}

\item{output}{Name of the output raster file.}

\item{wd}{Changes the working directory. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_wd()} for details.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_verbose()} for details.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_compress_rasters()} for details.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool. Default: \code{FALSE}.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a supervised parallelepiped classification using training site polygons and multi-spectral images.
}
\keyword{ImageProcessingToolsClassification}
