% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis_overlay.R
\name{wbt_percent_equal_to}
\alias{wbt_percent_equal_to}
\title{Percent equal to
Calculates the percentage of a raster stack that have cell values equal to an input on a cell-by-cell basis.}
\usage{
wbt_percent_equal_to(
  inputs,
  comparison,
  output,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{inputs}{Input raster files.}

\item{comparison}{Input comparison raster file.}

\item{output}{Output raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Percent equal to
Calculates the percentage of a raster stack that have cell values equal to an input on a cell-by-cell basis.
}
\keyword{GISAnalysisOverlayTools}
