% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{wbt_phi_coefficient}
\alias{wbt_phi_coefficient}
\title{Phi coefficient}
\usage{
wbt_phi_coefficient(
  input1,
  input2,
  output,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{input1}{Name of the first input raster image file.}

\item{input2}{Name of the second input raster image file.}

\item{output}{Name of the output HTML file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
This tool performs a binary classification accuracy assessment.
}
