% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis_enhancement.R
\name{wbt_piecewise_contrast_stretch}
\alias{wbt_piecewise_contrast_stretch}
\title{Piecewise contrast stretch}
\usage{
wbt_piecewise_contrast_stretch(
  input,
  output,
  FUN = "",
  greytones = 1024,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Name of the input raster image file.}

\item{output}{Name of the output raster image file.}

\item{FUN}{Piecewise break-points e.g. '(50, 0.1); (150, 0.8); (255; 1.0).}

\item{greytones}{Number of greytones in the output image.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a piecewise contrast stretch on an input image.
}
