% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_polygon_perimeter}
\alias{wbt_polygon_perimeter}
\title{Polygon perimeter}
\usage{
wbt_polygon_perimeter(input, wd = NULL, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input vector polygon file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the perimeter of vector polygons.
}
