% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{wbt_polygons_to_lines}
\alias{wbt_polygons_to_lines}
\title{Polygons to lines}
\usage{
wbt_polygons_to_lines(input, output, wd = NULL, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input vector polygon file.}

\item{output}{Output vector lines file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Converts vector polygons to polylines.
}
