% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_profile}
\alias{wbt_profile}
\title{Profile}
\usage{
wbt_profile(lines, surface, output, wd = NULL, verbose_mode = FALSE)
}
\arguments{
\item{lines}{Input vector line file.}

\item{surface}{Input raster surface file.}

\item{output}{Output HTML file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Plots profiles from digital surface models.
}
