% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{wbt_raise_walls}
\alias{wbt_raise_walls}
\title{Raise walls}
\usage{
wbt_raise_walls(
  input,
  dem,
  output,
  breach = NULL,
  height = 100,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input vector lines or polygons file.}

\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{breach}{Optional input vector breach lines.}

\item{height}{Wall height.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Raises walls in a DEM along a line or around a polygon, e.g. a watershed.
}
\keyword{HydrologicalAnalysis}
