% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{wbt_random_field}
\alias{wbt_random_field}
\title{Random field}
\usage{
wbt_random_field(base, output, wd = NULL, verbose_mode = FALSE)
}
\arguments{
\item{base}{Input raster file.}

\item{output}{Output raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Creates an image containing random values.
}
