% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_network_analysis.R
\name{wbt_raster_streams_to_vector}
\alias{wbt_raster_streams_to_vector}
\title{Raster streams to vector}
\usage{
wbt_raster_streams_to_vector(
  streams,
  d8_pntr,
  output,
  esri_pntr = FALSE,
  wd = NULL,
  verbose_mode = FALSE
)
}
\arguments{
\item{streams}{Input raster streams file.}

\item{d8_pntr}{Input raster D8 pointer file.}

\item{output}{Output vector file.}

\item{esri_pntr}{D8 pointer uses the ESRI style scheme.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Converts a raster stream file into a vector file.
}
