% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_network_analysis.R
\name{wbt_rasterize_streams}
\alias{wbt_rasterize_streams}
\title{Rasterize streams}
\usage{
wbt_rasterize_streams(
  streams,
  base,
  output,
  nodata = TRUE,
  feature_id = FALSE,
  wd = NULL,
  verbose_mode = NULL,
  compress_rasters = NULL,
  command_only = FALSE
)
}
\arguments{
\item{streams}{Input vector streams file.}

\item{base}{Input base raster file.}

\item{output}{Output raster file.}

\item{nodata}{Use NoData value for background?.}

\item{feature_id}{Use feature number as output value?.}

\item{wd}{Changes the working directory. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_wd()} for details.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_verbose()} for details.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_compress_rasters()} for details.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool. Default: \code{FALSE}.}
}
\value{
Returns the tool text outputs.
}
\description{
Rasterizes vector streams based on Lindsay (2016) method.
}
\keyword{StreamNetworkAnalysis}
