% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_relative_aspect}
\alias{wbt_relative_aspect}
\title{Relative aspect}
\usage{
wbt_relative_aspect(dem, output, azimuth = 0, zfactor = 1,
  verbose_mode = FALSE)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{azimuth}{Illumination source azimuth.}

\item{zfactor}{Optional multiplier for when the vertical and horizontal units are not the same.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates relative aspect (relative to a user-specified direction) from an input DEM.
}
