% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_remove_off_terrain_objects}
\alias{wbt_remove_off_terrain_objects}
\title{Remove off terrain objects}
\usage{
wbt_remove_off_terrain_objects(
  dem,
  output,
  filter = 11,
  slope = 15,
  verbose_mode = FALSE
)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{filter}{Filter size (cells).}

\item{slope}{Slope threshold value.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Removes off-terrain objects from a raster digital elevation model (DEM).
}
