% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{wbt_remove_spurs}
\alias{wbt_remove_spurs}
\title{Remove spurs}
\usage{
wbt_remove_spurs(
  input,
  output,
  iterations = 10,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{iterations}{Maximum number of iterations.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Removes the spurs (pruning operation) from a Boolean line image; intended to be used on the output of the LineThinning tool.
}
\keyword{ImageProcessingTools}
