% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_network_analysis.R
\name{wbt_repair_stream_vector_topology}
\alias{wbt_repair_stream_vector_topology}
\title{Repair stream vector topology}
\usage{
wbt_repair_stream_vector_topology(
  input,
  output,
  dist = "",
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{input}{Name of the input lines vector file.}

\item{output}{Name of the output lines vector file.}

\item{dist}{Snap distance, in xy units (metres).}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
This tool resolves topological errors and inconsistencies associated with digitized vector streams.
}
