% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{wbt_rgb_to_ihs}
\alias{wbt_rgb_to_ihs}
\title{Rgb to ihs}
\usage{
wbt_rgb_to_ihs(
  intensity,
  hue,
  saturation,
  red = NULL,
  green = NULL,
  blue = NULL,
  composite = NULL,
  verbose_mode = FALSE
)
}
\arguments{
\item{intensity}{Output intensity raster file.}

\item{hue}{Output hue raster file.}

\item{saturation}{Output saturation raster file.}

\item{red}{Input red band image file. Optionally specified if colour-composite not specified.}

\item{green}{Input green band image file. Optionally specified if colour-composite not specified.}

\item{blue}{Input blue band image file. Optionally specified if colour-composite not specified.}

\item{composite}{Input colour-composite image file. Only used if individual bands are not specified.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Converts red, green, and blue (RGB) images into intensity, hue, and saturation (IHS) images.
}
