% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{wbt_roberts_cross_filter}
\alias{wbt_roberts_cross_filter}
\title{Roberts cross filter}
\usage{
wbt_roberts_cross_filter(input, output, clip = 0, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{clip}{Optional amount to clip the distribution tails by, in percent.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a Robert's cross edge-detection filter on an image.
}
