% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{wbt_root_mean_square_error}
\alias{wbt_root_mean_square_error}
\title{Root mean square error}
\usage{
wbt_root_mean_square_error(input, base, wd = NULL, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{base}{Input base raster file used for comparison.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the RMSE and other accuracy statistics.
}
