% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_select_tiles_by_polygon}
\alias{wbt_select_tiles_by_polygon}
\title{Select tiles by polygon}
\usage{
wbt_select_tiles_by_polygon(
  indir,
  outdir,
  polygons,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{indir}{Input LAS file source directory.}

\item{outdir}{Output directory into which LAS files within the polygon are copied.}

\item{polygons}{Input vector polygons file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Copies LiDAR tiles overlapping with a polygon into an output directory.
}
