% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis_enhancement.R
\name{wbt_sigmoidal_contrast_stretch}
\alias{wbt_sigmoidal_contrast_stretch}
\title{Sigmoidal contrast stretch}
\usage{
wbt_sigmoidal_contrast_stretch(
  input,
  output,
  cutoff = 0,
  gain = 1,
  num_tones = 256,
  wd = NULL,
  verbose_mode = NULL,
  compress_rasters = NULL,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{cutoff}{Cutoff value between 0.0 and 0.95.}

\item{gain}{Gain value.}

\item{num_tones}{Number of tones in the output image.}

\item{wd}{Changes the working directory. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_wd()} for details.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_verbose()} for details.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_compress_rasters()} for details.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool. Default: \code{FALSE}.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a sigmoidal contrast stretch on input images.
}
\keyword{ImageProcessingToolsImageEnhancement}
