% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{wbt_sigmoidal_contrast_stretch}
\alias{wbt_sigmoidal_contrast_stretch}
\title{Sigmoidal contrast stretch}
\usage{
wbt_sigmoidal_contrast_stretch(
  input,
  output,
  cutoff = 0,
  gain = 1,
  num_tones = 256,
  wd = NULL,
  verbose_mode = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{cutoff}{Cutoff value between 0.0 and 0.95.}

\item{gain}{Gain value.}

\item{num_tones}{Number of tones in the output image.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a sigmoidal contrast stretch on input images.
}
