% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{wbt_single_part_to_multi_part}
\alias{wbt_single_part_to_multi_part}
\title{Single part to multi part}
\usage{
wbt_single_part_to_multi_part(
  input,
  output,
  field = NULL,
  wd = NULL,
  verbose_mode = FALSE
)
}
\arguments{
\item{input}{Input vector line or polygon file.}

\item{output}{Output vector line or polygon file.}

\item{field}{Grouping ID field name in attribute table.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Converts a vector file containing multi-part features into a vector containing only single-part features.
}
