% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_slope}
\alias{wbt_slope}
\title{Slope
Calculates a slope raster from an input DEM.}
\usage{
wbt_slope(
  dem,
  output,
  zfactor = NULL,
  units = "degrees",
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{zfactor}{Optional multiplier for when the vertical and horizontal units are not the same.}

\item{units}{Units of output raster; options include 'degrees', 'radians', 'percent'.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Slope
Calculates a slope raster from an input DEM.
}
\keyword{GeomorphometricAnalysis}
