% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_smooth_vectors}
\alias{wbt_smooth_vectors}
\title{Smooth vectors}
\usage{
wbt_smooth_vectors(input, output, filter = 3, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input vector POLYLINE or POLYGON file.}

\item{output}{Output vector file.}

\item{filter}{The filter size, any odd integer greater than or equal to 3; default is 3.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Smooths a vector coverage of either a POLYLINE or POLYGON base ShapeType.
}
