% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{wbt_snap_pour_points}
\alias{wbt_snap_pour_points}
\title{Snap pour points}
\usage{
wbt_snap_pour_points(pour_pts, flow_accum, output, snap_dist,
  verbose_mode = FALSE)
}
\arguments{
\item{pour_pts}{Input vector pour points (outlet) file.}

\item{flow_accum}{Input raster D8 flow accumulation file.}

\item{output}{Output vector file.}

\item{snap_dist}{Maximum snap distance in map units.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Moves outlet points used to specify points of interest in a watershedding operation to the cell with the highest flow accumulation in its neighbourhood.
}
