% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis_filters.R
\name{wbt_sobel_filter}
\alias{wbt_sobel_filter}
\title{Sobel filter}
\usage{
wbt_sobel_filter(
  input,
  output,
  variant = "3x3",
  clip = 0,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{variant}{Optional variant value. Options include 3x3 and 5x5 (default is 3x3).}

\item{clip}{Optional amount to clip the distribution tails by, in percent (default is 0.0).}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Performs a Sobel edge-detection filter on an image.
}
\keyword{ImageProcessingToolsFilters}
