% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{wbt_split_colour_composite}
\alias{wbt_split_colour_composite}
\title{Split colour composite
This tool splits an RGB colour composite image into separate multispectral images.}
\usage{
wbt_split_colour_composite(
  input,
  red = NULL,
  green = NULL,
  blue = NULL,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Input colour composite image file.}

\item{red}{Output red band file.}

\item{green}{Output green band file.}

\item{blue}{Output blue band file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Split colour composite
This tool splits an RGB colour composite image into separate multispectral images.
}
\keyword{ImageProcessingTools}
