% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{wbt_split_colour_composite}
\alias{wbt_split_colour_composite}
\title{Split colour composite}
\usage{
wbt_split_colour_composite(
  input,
  red = NULL,
  green = NULL,
  blue = NULL,
  verbose_mode = FALSE
)
}
\arguments{
\item{input}{Input colour composite image file.}

\item{red}{Output red band file.}

\item{green}{Output green band file.}

\item{blue}{Output blue band file.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
This tool splits an RGB colour composite image into seperate multispectral images.
}
