% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_split_vector_lines}
\alias{wbt_split_vector_lines}
\title{Split vector lines}
\usage{
wbt_split_vector_lines(
  input,
  output,
  length = NULL,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Name of the input lines shapefile.}

\item{output}{Name of the output lines shapefile.}

\item{length}{Maximum segment length (m).}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
This tool can be used to split a vector line coverage into even-lengthed segments.
}
