% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbt.R
\name{wbt_tool_help}
\alias{wbt_tool_help}
\title{Help description for a specific tool.}
\usage{
wbt_tool_help(tool_name = NULL)
}
\arguments{
\item{tool_name}{The name of the tool.}
}
\value{
Returns the help description for a specific tool.
}
\description{
Retrieves the help description for a specific tool.
}
\examples{
\dontrun{
wbt_tool_help("lidar_info")
}
}
