% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_stat_analysis.R
\name{wbt_turning_bands_simulation}
\alias{wbt_turning_bands_simulation}
\title{Turning bands simulation}
\usage{
wbt_turning_bands_simulation(
  base,
  output,
  range,
  iterations = 1000,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{base}{Input base raster file.}

\item{output}{Output file.}

\item{range}{The field's range, in xy-units, related to the extent of spatial autocorrelation.}

\item{iterations}{The number of iterations.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Creates an image containing random values based on a turning-bands simulation.
}
