% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis.R
\name{wbt_union}
\alias{wbt_union}
\title{Union}
\usage{
wbt_union(input, overlay, output, snap = 0, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input vector file.}

\item{overlay}{Input overlay vector file.}

\item{output}{Output vector file.}

\item{snap}{Snap tolerance.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Splits vector layers at their overlaps, creating a layer containing all the portions from both input and overlay layers.
}
