% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{wbt_unsharp_masking}
\alias{wbt_unsharp_masking}
\title{Unsharp masking}
\usage{
wbt_unsharp_masking(input, output, sigma = 0.75, amount = 100,
  threshold = 0, verbose_mode = FALSE)
}
\arguments{
\item{input}{Input raster file.}

\item{output}{Output raster file.}

\item{sigma}{Standard deviation distance in pixels.}

\item{amount}{A percentage and controls the magnitude of each overshoot.}

\item{threshold}{Controls the minimal brightness change that will be sharpened.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
An image sharpening technique that enhances edges.
}
