% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_analysis_overlay.R
\name{wbt_update_nodata_cells}
\alias{wbt_update_nodata_cells}
\title{Update nodata cells}
\usage{
wbt_update_nodata_cells(
  input1,
  input2,
  output,
  wd = NULL,
  verbose_mode = NULL,
  compress_rasters = NULL,
  command_only = FALSE
)
}
\arguments{
\item{input1}{Input raster file 1.}

\item{input2}{Input raster file 2; update layer.}

\item{output}{Output raster file.}

\item{wd}{Changes the working directory. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_wd()} for details.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_verbose()} for details.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters. Default: \code{NULL} will use the value in WhiteboxTools settings, see \code{wbt_compress_rasters()} for details.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool. Default: \code{FALSE}.}
}
\value{
Returns the tool text outputs.
}
\description{
Replaces the NoData values in an input raster with the corresponding values contained in a second update layer.
}
\keyword{GISAnalysisOverlayTools}
