% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbt.R
\name{wbt_verbose}
\alias{wbt_verbose}
\title{Check verbose options for WhiteboxTools.}
\usage{
wbt_verbose()
}
\value{
logical; defaults to result of `interactive()`
}
\description{
Return `TRUE` if "R_WHITEBOX_VERBOSE" system environment variable or "whitebox.verbose" option is set to `TRUE`. Returns `TRUE` when R is being used interactively and `FALSE` otherwise.
}
\examples{
 wbt_verbose()
 
\dontrun{
 Sys.setenv("R_WHITEBOX_VERBOSE" = FALSE)
 wbt_verbose()
 
 options(whitebox.verbose = FALSE)
 wbt_verbose()
}
}
