% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbt.R
\name{wbt_version}
\alias{wbt_version}
\title{Version information for WhiteboxTools}
\usage{
wbt_version(extract = FALSE)
}
\arguments{
\item{extract}{Extract semantic version number from first line of result? Default: \code{FALSE}}
}
\value{
Returns the version information for WhiteboxTools as an R character vector.
}
\description{
Version information for WhiteboxTools
}
\examples{
\dontrun{
wbt_version()
}
}
