% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_visibility_index}
\alias{wbt_visibility_index}
\title{Visibility index}
\usage{
wbt_visibility_index(
  dem,
  output,
  height = 2,
  res_factor = 2,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{dem}{Input raster DEM file.}

\item{output}{Output raster file.}

\item{height}{Viewing station height, in z units.}

\item{res_factor}{The resolution factor determines the density of measured viewsheds.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
Estimates the relative visibility of sites in a DEM.
}
