% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_analysis.R
\name{wbt_watershed}
\alias{wbt_watershed}
\title{Watershed}
\usage{
wbt_watershed(
  d8_pntr,
  pour_pts,
  output,
  esri_pntr = FALSE,
  verbose_mode = FALSE
)
}
\arguments{
\item{d8_pntr}{Input D8 pointer raster file.}

\item{pour_pts}{Input pour points (outlet) file.}

\item{output}{Output raster file.}

\item{esri_pntr}{D8 pointer uses the ESRI style scheme.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}
}
\value{
Returns the tool text outputs.
}
\description{
Identifies the watershed, or drainage basin, draining to a set of target cells.
}
