% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{wbt_wetness_index}
\alias{wbt_wetness_index}
\title{Wetness index}
\usage{
wbt_wetness_index(
  sca,
  slope,
  output,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{sca}{Input raster specific contributing area (SCA) file.}

\item{slope}{Input raster slope file (in degrees).}

\item{output}{Output raster file.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Calculates the topographic wetness index, Ln(A / tan(slope)).
}
\keyword{GeomorphometricAnalysis}
