% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_agriculture.R
\name{wbt_yield_filter}
\alias{wbt_yield_filter}
\title{Yield filter}
\usage{
wbt_yield_filter(
  input,
  yield_field_name,
  pass_field_name,
  output,
  width = 6.096,
  z_score_threshold = 2.5,
  min_yield = 0,
  max_yield = 99999.9,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{input}{Name of the input points shapefile.}

\item{yield_field_name}{Name of the attribute containing yield data.}

\item{pass_field_name}{Name of the attribute containing pass line ID.}

\item{output}{Name of the output points shapefile.}

\item{width}{Pass swath width (m).}

\item{z_score_threshold}{Z-score threshold value (default=2.5).}

\item{min_yield}{Minimum yield value in output.}

\item{max_yield}{Maximum yield value in output.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
This tool can be used to approximate the harvester pass lines from yield points.
}
