% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_agriculture.R
\name{wbt_yield_map}
\alias{wbt_yield_map}
\title{Yield map
This tool can be used to create a segmented-vector polygon yield map from a set of harvester points.}
\usage{
wbt_yield_map(
  input,
  pass_field_name,
  output,
  width = 6.096,
  max_change_in_heading = 25,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE,
  command_only = FALSE
)
}
\arguments{
\item{input}{Name of the input points shapefile.}

\item{pass_field_name}{Name of the attribute containing pass line ID.}

\item{output}{Name of the output polygon shapefile.}

\item{width}{Pass swath width (m).}

\item{max_change_in_heading}{Max change in heading.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is \code{FALSE}, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters.}

\item{command_only}{Return command that would be executed by \code{system()} rather than running tool.}
}
\value{
Returns the tool text outputs.
}
\description{
Yield map
This tool can be used to create a segmented-vector polygon yield map from a set of harvester points.
}
\keyword{PrecisionAgriculture}
