% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_analysis.R
\name{wbt_zlidar_to_las}
\alias{wbt_zlidar_to_las}
\title{zlidar to las}
\usage{
wbt_zlidar_to_las(
  inputs = NULL,
  outdir = NULL,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)
}
\arguments{
\item{inputs}{Input zlidar files.}

\item{outdir}{Output directory into which zlidar files are created. If unspecified, it is assumed to be the same as the inputs.}

\item{wd}{Changes the working directory.}

\item{verbose_mode}{Sets verbose mode. If verbose mode is False, tools will not print output messages.}

\item{compress_rasters}{Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.}
}
\value{
Returns the tool text outputs.
}
\description{
Converts one or more zlidar files into the LAS data format.
}
