% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whitebox-package.R
\docType{data}
\name{wbttoolparameters}
\alias{wbttoolparameters}
\title{'WhiteboxTools' Tool Parameters}
\format{
A \code{data.frame} with 2169 observations of 13 variables
\itemize{
\item \code{"function_name"} - R function name
\item \code{"tool_name"} - WhiteboxTools tool name
\item \code{"name"} - parameter name
\item \code{"flags"} - flags used to specify parameter on command line; comma separated
\item \code{"description"} - parameter description
\item \code{"parameter_class"} - parameter type
\item \code{"parameter_detail"} - parameter details; character: data type followed by colon and more specifics, For OptionList possible values, comma-separated (if defined)
\item \code{"default_value"} - parameter default value, if any
\item \code{"optional"} - parameter "optional" flag; note that some combination of optional parameters may be required for certain conditions
\item \code{"label"} - labels for selected subset of \code{"flags"} \strong{used as R function argument names} for \code{wbt_} functions
\item \code{"is_input"} - logical. Classification of 'input' parameters
\item \code{"is_output"} - logical. Classification of 'output' parameters
}
}
\source{
\href{https://github.com/jblindsay/whitebox-tools/releases/}{WhiteboxTools}
}
\usage{
wbttoolparameters
}
\description{
This data set is a \code{data.frame} containing tool parameters and associated metadata
}
\seealso{
\link{wbttools} \code{\link[=wbt_tool_parameters]{wbt_tool_parameters()}}
}
\keyword{datasets}
