% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whitebox-package.R
\docType{data}
\name{wbttoolparameters}
\alias{wbttoolparameters}
\title{WhiteboxTools Tool Parameters}
\format{
A \code{data.frame} with 1706 observations of 13 variables
\itemize{
\item \code{"tool_name"} - WhiteboxTools tool name
\item \code{"name"} - parameter name
\item \code{"flags"} - flags used to specify parameter on command line; comma separated
\item \code{"description"} - parameter description
\item \code{"parameter_type"} - parameter data type; character: data type(s), space, then possible values, comma-separated (if defined)
\item \code{"default_value"} - parameter default value, if any
\item \code{"optional"} - parameter "optional" flag; note that some combination of optional parameters may be required for certain conditions
\item \code{"pkgflags"} - selected flags 1:1 with package function argument names (notably: \code{-i} and \code{-o} omitted)
\item \code{"label"} - labels for \code{"pkgflags"} used as argument names
\item \code{"is_input"} - logical. Classification of 'input' parameters
\item \code{"is_output"} - logical. Classification of 'output' parameters
\item \code{"function"} - R function name
}
}
\source{
\href{https://github.com/jblindsay/whitebox-tools/releases/tag/1.5.0}{WhiteboxTools 1.5.0}
}
\description{
This data set is a \code{data.frame} containing tools by name, their parameters, and associated metadata, as available in WhiteboxTools 1.5.0
}
\seealso{
\link{wbttools} \code{\link[=wbt_tool_parameters]{wbt_tool_parameters()}}
}
\keyword{datasets}
