% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whitebox-package.R
\docType{data}
\name{wbttools}
\alias{wbttools}
\title{WhiteboxTools Tool List}
\format{
A data.frame with 448 observations of 4 variables
\itemize{
\item \code{"tool_name"} - WhiteboxTools tool name
\item \code{"toolbox_name"} - WhiteboxTools toolbox name
\item \code{"description"} - Brief description
\item \code{"function_name"} - R function name
}
}
\source{
\href{https://github.com/jblindsay/whitebox-tools/releases/tag/1.5.0}{WhiteboxTools 1.5.0}
}
\description{
This data set is a \code{data.frame} containing tools by name and associated R function name, as available in WhiteboxTools 1.5.0
}
\keyword{datasets}
