% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whitebox-package.R
\docType{data}
\name{wbttools}
\alias{wbttools}
\title{WhiteboxTools Tool List}
\format{
A data.frame with 518 observations of 7 variables
\itemize{
\item \code{"tool_name"} - WhiteboxTools tool name
\item \code{"function_name"} - R function name
\item \code{"toolbox_name"} - WhiteboxTools toolbox name
\item \code{"label"} - WhiteboxTools tool label
\item \code{"description"} - Brief description
\item \code{"github"} - Link to related code on GitHub
\item \code{"book"} - Link to WhiteboxTools Manual
}
}
\source{
\href{https://github.com/jblindsay/whitebox-tools/releases/}{WhiteboxTools}
}
\usage{
wbttools
}
\description{
This data set is a \code{data.frame} containing tools by name and associated R function name
}
\seealso{
\link{wbttoolparameters} \code{\link[=wbt_list_tools]{wbt_list_tools()}}
}
\keyword{datasets}
